==> code/03_beta_diversity.Rmd <==
---
title: "Beta diversity"
author: "Haipeng Sun"
date: "7/14/2022"
output: html_document
---

## library
```{r setup, message=FALSE, warning=FALSE}
library(knitr)
library(phyloseq)
library(usedist)
library(purrr)
library(furrr)
library(vegan)
library(ape)
library(rlist)
library(ggpubr)
library(ggtext)
library(ggrepel)
library(glue)
library(patchwork)
library(gridExtra)
library(gtable)
library(tidyverse)
```

## load data
```{r}
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/phylo_meta.Rdata")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_uu.RData")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_wu.RData")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_bray.RData")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_jaccard.RData")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/alpha.Rdata")
temp_out = "~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/output/beta/"
```

## function
```{r}
calcOmega2_dm <- function(x) {
    # Computes the effect size omega^2 parital using the output from adonis test
    # Args:
    #   x: adonis output
    # Returns:
    #   A dataframe with calculated effect size omega^2 partial added to original dataframe
    require(dplyr)
    N_t = x["Total", ]$Df + 1
    MSe = x["Residual", ]$SumOfSqs/x["Residual", ]$Df
    out = x %>% as.data.frame %>% rownames_to_column(var = "Variable") %>% mutate(Omega2_partial = ifelse(is.na(`F`), NA, (SumOfSqs - Df * MSe)/(SumOfSqs + (N_t - Df) * MSe)))
    return(out)
}
```

## beta explore all 
```{r}
for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta = meta_all %>% filter(SampleID %in% labels(wdist), !is.na(Country))
  wdist <- dist_subset(wdist, wmeta$SampleID)  
  ## dispersion of major variables
  dp <- map_dfr(c("Country","Continent","Extraction_method","heat_inactivated","Project"), 
                function(a){
                  disper <- betadisper(wdist, wmeta %>% pull(a))
                  res <- (TukeyHSD(disper, which = "group"))$group %>% 
                    as.data.frame() %>% rownames_to_column(var = "Compare")
                  res$Variable = a
                  res
                  }, .id = NULL)
  write.csv(dp, file = paste0(temp_out,"dispersion_", dd,"_overall.csv"), quote = F, row.names = F)
  ## adonis of major variables country
  per <- adonis2(wdist ~ PCRConc+heat_inactivated+Extraction_method+Country, 
                 data = wmeta, by = "margin")
  res = calcOmega2_dm(per)
  write.csv(res, 
            file = paste0(temp_out,"adonis_test_", dd, "_overall_Country.csv"), quote = F, row.names = F)
  per <- adonis2(wdist ~ PCRConc+heat_inactivated+Extraction_method+Continent, 
                 data = wmeta, by = "margin")
  res = calcOmega2_dm(per)
  write.csv(res, 
            file = paste0(temp_out,"adonis_test_", dd, "_overall_Continent.csv"), quote = F, row.names = F)  
}
```

## beta in Global project
```{r}
meta <- phylo_decontam_clean_rf@sam_data@.Data %>% as.data.frame()
names(meta) <- phylo_decontam_clean_rf@sam_data@names
meta$SampleID <- phylo_decontam_clean_rf@sam_data@row.names

meta <- rbind(meta %>% filter(!is.na(Country), Project %in% c("Global", "Upside", "Milkbank")),
             meta %>% filter(!is.na(Country), Project=="PR") %>%
               filter(TimePoint == 1200))

for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta = meta %>% filter(SampleID %in% labels(wdist))
  wdist <- dist_subset(wdist, wmeta$SampleID) 
  #adonis
  # per <- adonis2(wdist ~ Country, 
  #                data = wmeta, by = "margin")
  # res = calcOmega2_dm(per)
  # write.csv(res, 
  #           file = paste0(temp_out,"adonis_test_", dd, "_Global_Country_only.csv"), quote = F, row.names = F)
  # per <- adonis2(wdist ~ Continent, 
  #                data = wmeta, by = "margin")
  # res = calcOmega2_dm(per)
  # write.csv(res, 
  #           file = paste0(temp_out,"adonis_test_", dd, "_Global_Continent_only.csv"), quote = F, row.names = F) 
  # per <- adonis2(wdist ~ Conti_grp, 
  #                data = wmeta, by = "margin")
  # res = calcOmega2_dm(per)
  # write.csv(res, 
  #           file = paste0(temp_out,"adonis_test_", dd, "_Global_Continent_group.csv"), quote = F, row.names = F)   
  #PCoA
  disper <- betadisper(wdist, wmeta$Country)   
  anova(disper)
  disper <- betadisper(wdist, wmeta$Continent)   
  anova(disper)
  
  
  wrda <- dbrda(wdist ~ 1, sqrt.dist = TRUE)
  wrda_eig <- eigenvals(wrda)
  varPC1 <- round(wrda_eig["MDS1"]/sum(wrda_eig)*100,2)
  varPC2 <- round(wrda_eig["MDS2"]/sum(wrda_eig)*100,2)
  x_lab <- paste0("PC1 (", varPC1, "%)")
  y_lab <- paste0("PC2 (", varPC2, "%)")
  wdat_pcoa <- merge(scores(wrda, display = "sites", choices = c(1,2)),
                    wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = MDS1, PC2 = MDS2)
  ## Country
  wrda_fit <- envfit(wrda ~ Country, data = wmeta)
  wdat_center <- wrda_fit$factors$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "Country") %>%
    transmute(Country = str_remove(Country,"Country"),PC1 = MDS1, PC2 = MDS2) %>%
    left_join(wmeta %>% dplyr::select(Country, Continent) %>% unique(),by = "Country")
  wdat <- wdat_pcoa %>% left_join(wdat_center %>% transmute(PC1.c = PC1, PC2.c = PC2, Country), by = "Country")
  g_tab <- read.csv(paste0(temp_out,"adonis_test_", dd, "_Global_Country_only.csv")) %>% 
    filter(!is.na(Omega2_partial)) %>% 
    transmute(Name = Variable, Df = Df, SumOfSqs = round(SumOfSqs, 3),
              R2 = round(R2, 3), `F` = round(`F`, 3), Pvalue = round(`Pr..F.`,3),
              Omega2_partial = round(Omega2_partial,3)) %>% column_to_rownames(var = "Name")
  g <- ggplot(wdat, aes(x = PC1, y = PC2, color = Country)) + 
      geom_point(size = 2) + 
      scale_color_manual(values = Country_color) + 
      labs(x = x_lab, y = y_lab, title = "") +
      theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin)   
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_Global_by_Country.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)
  
  g <- ggplot(wdat_center, aes(x = PC1, y = PC2, color = Continent)) + 
    geom_point(data = wdat, aes(x = PC1, y = PC2, color = Continent), size = 1, alpha = 0.5) + 
    geom_segment(data = wdat, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4,shape = 21, color = "black", aes(fill = Continent)) + 
    geom_text_repel(aes(label = Country), color = "black") + 
    scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold')) 
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin)    
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_Global_by_Country_center.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)  
  ## Continent center
  wrda_fit <- envfit(wrda ~ Continent, data = wmeta)
  wdat_center2 <- wrda_fit$factors$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "Continent") %>%
    transmute(Continent = str_remove(Continent,"Continent"),PC1 = MDS1, PC2 = MDS2)
  wdat2 <- wdat_pcoa %>% left_join(wdat_center2 %>% transmute(PC1.c = PC1, PC2.c = PC2, Continent), by = "Continent")
  g_tab <- read.csv(paste0(temp_out,"adonis_test_", dd, "_Global_Continent_only.csv")) %>% 
    filter(!is.na(Omega2_partial)) %>% 
    transmute(Name = Variable, Df = Df, SumOfSqs = round(SumOfSqs, 3),
              R2 = round(R2, 3), `F` = round(`F`, 3), Pvalue = round(`Pr..F.`,3),
              Omega2_partial = round(Omega2_partial,3)) %>% column_to_rownames(var = "Name")
  g <- ggplot(wdat_center2, aes(x = PC1, y = PC2, color = Continent)) + 
    geom_point(data = wdat2, aes(x = PC1, y = PC2, color = Continent), size = 1, alpha = 0.5) + 
    geom_segment(data = wdat2, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4, shape = 21, color = "black", aes(fill = Continent)) + 
    geom_text_repel(aes(label = Continent), color = "black") + 
    scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
    #stat_ellipse(data = wdat2, aes(x = PC1, y = PC2, color = Continent)) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin)   
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_Global_by_Continent_center.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)  

  ## Continent group center
  wrda_fit <- envfit(wrda ~ Conti_grp, data = wmeta)
  wdat_center2 <- wrda_fit$factors$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "Conti_grp") %>%
    transmute(Conti_grp = str_remove(Conti_grp,"Conti_grp"),PC1 = MDS1, PC2 = MDS2)
  wdat2 <- wdat_pcoa %>% left_join(wdat_center2 %>% transmute(PC1.c = PC1, PC2.c = PC2, Conti_grp), by = "Conti_grp")
  g_tab <- read.csv(paste0(temp_out,"adonis_test_", dd, "_Global_Continent_group.csv")) %>% 
    filter(!is.na(Omega2_partial)) %>% 
    transmute(Name = Variable, Df = Df, SumOfSqs = round(SumOfSqs, 3),
              R2 = round(R2, 3), `F` = round(`F`, 3), Pvalue = round(`Pr..F.`,3),
              Omega2_partial = round(Omega2_partial,3)) %>% column_to_rownames(var = "Name")  
  g <- ggplot(wdat_center2, aes(x = PC1, y = PC2, color = Conti_grp)) + 
    geom_point(data = wdat2, aes(x = PC1, y = PC2, color = Conti_grp), size = 1, alpha = 0.5) + 
    geom_segment(data = wdat2, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4, shape = 21, color = "black", aes(fill = Conti_grp)) + 
    geom_text_repel(aes(label = Conti_grp), color = "black") + 
    scale_color_manual(values = c("Africa"="#7fc97f", "Europe"="#fdc086",
                                  "Central and South America"="#6A3D9A",
                                  "United States"="#BEAED4","Puerto Rico"="#8DA0CB"),
                       aesthetics = c("color","fill"),
                       labels = c("Africa","Europe",
                                  "C&South America",
                                  "United States","Puerto Rico")) + 
    #stat_ellipse(data = wdat2, aes(x = PC1, y = PC2, color = Continent)) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))   
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin)    
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_Global_by_Continent_group.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)    
  
  #RDA Country
  wrda <- dbrda(wdist ~ Country, sqrt.dist = TRUE, data = wmeta)
  wrda_eig <- eigenvals(wrda)
  varRDA1 <- round(wrda_eig["dbRDA1"]/sum(wrda_eig)*100,2)
  varRDA2 <- round(wrda_eig["dbRDA2"]/sum(wrda_eig)*100,2)
  wrda_score <- scores(wrda)
  wdat_rda <- merge(wrda_score$sites,
                    wmeta, by.x = 0, by.y = "SampleID", all.x = T)
  wdat_center <- wrda_score$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "Country") %>%
    transmute(Country = str_remove(Country,"Country"),dbRDA1, dbRDA2) %>%
    left_join(wmeta %>% dplyr::select(Country, Continent) %>% unique(),by = "Country")
  wdat <- wdat_rda %>% left_join(wdat_center %>% transmute(dbRDA1.c = dbRDA1, dbRDA2.c = dbRDA2, Country), by = "Country")
  x_lab <- paste0("dbRDA1 (", varRDA1, "%)")
  y_lab <- paste0("dbRDA2 (", varRDA2, "%)")  
  
  g <- ggplot(wdat_center, aes(x = dbRDA1, y = dbRDA2, color = Continent)) + 
    geom_point(data = wdat, aes(x = dbRDA1, y = dbRDA2, color = Continent), size = 1, alpha = 0.5) + 
    geom_segment(data = wdat, aes(x = dbRDA1, y = dbRDA2, xend = dbRDA1.c, yend = dbRDA2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4, shape = 21, color = "black", aes(fill = Continent)) + 
    geom_text_repel(aes(label = Country), color = "black") + 
    scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))    
  ggsave(filename = paste0(temp_out,"dbRDA_",dd,"_Global_by_Country_center.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)
  
  #RDA Continent
  wrda <- dbrda(wdist ~ Continent, sqrt.dist = TRUE, data = wmeta)
  wrda_eig <- eigenvals(wrda)
  varRDA1 <- round(wrda_eig["dbRDA1"]/sum(wrda_eig)*100,2)
  varRDA2 <- round(wrda_eig["dbRDA2"]/sum(wrda_eig)*100,2)
  x_lab <- paste0("dbRDA1 (", varRDA1, "%)")
  y_lab <- paste0("dbRDA2 (", varRDA2, "%)")  
  wrda_score <- scores(wrda)
  wdat_rda <- merge(wrda_score$sites,
                    wmeta, by.x = 0, by.y = "SampleID", all.x = T)
  wdat_center <- wrda_score$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "Continent") %>%
    mutate(Continent = str_remove(Continent,"Continent"))
  wdat <- wdat_rda %>% left_join(wdat_center %>% transmute(dbRDA1.c = dbRDA1, dbRDA2.c = dbRDA2, Continent), by = "Continent")

  g <- ggplot(wdat_center, aes(x = dbRDA1, y = dbRDA2, color = Continent)) + 
    geom_point(data = wdat, aes(x = dbRDA1, y = dbRDA2, color = Continent), size = 1, alpha = 0.5) + 
    geom_segment(data = wdat, aes(x = dbRDA1, y = dbRDA2, xend = dbRDA1.c, yend = dbRDA2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4, shape = 21, color = "black", aes(fill = Continent)) + 
    geom_text_repel(aes(label = Continent), color = "black") + 
    scale_color_manual(values = Continent_color,aesthetics = c("color","fill")) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))    
  ggsave(filename = paste0(temp_out,"dbRDA_",dd,"_Global_by_Continent_center.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)
}
```

##
```{r}
for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta = meta %>% filter(SampleID %in% labels(wdist))
  wdist <- dist_subset(wdist, wmeta$SampleID) 
  print(dd)
  disper <- betadisper(wdist, wmeta$Country)   
  anova(disper) %>% print()
  disper <- betadisper(wdist, wmeta$Continent)   
  anova(disper) %>% print()
}
```



## beta diversity by alpha
```{r}
dat_alpha <- left_join(alpha_boot, meta_all, by = "SampleID")
alpha <- rbind(dat_alpha %>% filter(!is.na(Country), Project %in% c("Global", "Upside", "Milkbank")),
               dat_alpha %>% filter(!is.na(Country), Project=="PR") %>% filter(TimePoint == 1200))
#hist(alpha$Faith_PD)
alpha_cutoff <- 10
meta<- alpha %>% mutate(alpha_grp1 = ifelse(Faith_PD > alpha_cutoff, ">10", "<=10"),
                          alpha_grp2 = ntile(Faith_PD, 3)) %>% 
  mutate(alpha_grp2 = case_when(alpha_grp2 == 1 ~ "Low",
                                alpha_grp2 == 2 ~ "Medium",
                                alpha_grp2 == 3 ~ "High"))

alpha_grp1_color = c("<=10"="#d53e4f",">10"="#3288bd")
alpha_grp2_color = c("Low"="#d53e4f", "Medium"="#bababa","High"="#3288bd")

for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta = meta %>% filter(SampleID %in% labels(wdist))
  wdist <- dist_subset(wdist, wmeta$SampleID) 
  #adonis
  #per <- adonis2(wdist ~ alpha_grp1 + Country, 
  #               data = wmeta, by = "margin")
  #res = calcOmega2_dm(per)
  #write.csv(res, 
  #          file = paste0(temp_out,"adonis_test_", dd, "_Global_Country_alpha1.csv"), quote = F, row.names = F)
  #per <- adonis2(wdist ~ alpha_grp1 + Continent, 
  #               data = wmeta, by = "margin")
  #res = calcOmega2_dm(per)
  #write.csv(res, 
  #          file = paste0(temp_out,"adonis_test_", dd, "_Global_Continent_alpha1.csv"), quote = F, row.names = F) 
  #per <- adonis2(wdist ~ alpha_grp2 + Country, 
  #               data = wmeta, by = "margin")
  #res = calcOmega2_dm(per)
  #write.csv(res, 
  #          file = paste0(temp_out,"adonis_test_", dd, "_Global_Country_alpha2.csv"), quote = F, row.names = F)
  #per <- adonis2(wdist ~ alpha_grp2 + Continent, 
  #               data = wmeta, by = "margin")
  #res = calcOmega2_dm(per)
  #write.csv(res, 
  #          file = paste0(temp_out,"adonis_test_", dd, "_Global_Continent_alpha2.csv"), quote = F, row.names = F) 
  #per <- adonis2(wdist ~ alpha_grp1, data = wmeta, by = "margin")
  #res = calcOmega2_dm(per)
  #write.csv(res, 
  #          file = paste0(temp_out,"adonis_test_", dd, "_Global_alpha1_only.csv"), quote = F, row.names = F)
  #per <- adonis2(wdist ~ alpha_grp2, data = wmeta, by = "margin")
  #res = calcOmega2_dm(per)
  #write.csv(res, 
  #          file = paste0(temp_out,"adonis_test_", dd, "_Global_alpha2_only.csv"), quote = F, row.names = F) 
  #PCoA
  wrda <- dbrda(wdist ~ 1, sqrt.dist = TRUE)
  wrda_eig <- eigenvals(wrda)
  varPC1 <- round(wrda_eig["MDS1"]/sum(wrda_eig)*100,2)
  varPC2 <- round(wrda_eig["MDS2"]/sum(wrda_eig)*100,2)
  x_lab <- paste0("PC1 (", varPC1, "%)")
  y_lab <- paste0("PC2 (", varPC2, "%)")  
  wdat_pcoa <- merge(scores(wrda, display = "sites", choices = c(1,2)),
                    wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = MDS1, PC2 = MDS2)
  #alpha grp 1
  wrda_fit <- envfit(wrda ~ alpha_grp1, data = wmeta)
  wdat_center <- wrda_fit$factors$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "alpha_grp1") %>%
    transmute(alpha_grp1 = str_remove(alpha_grp1,"alpha_grp1"),PC1 = MDS1, PC2 = MDS2) 
  wdat <- wdat_pcoa %>% left_join(wdat_center %>% transmute(PC1.c = PC1, PC2.c = PC2, alpha_grp1), by = "alpha_grp1")
  g_tab <- read.csv(paste0(temp_out,"adonis_test_", dd, "_Global_alpha1_only.csv")) %>% 
    filter(!is.na(Omega2_partial)) %>% 
    transmute(Name = Variable, Df = Df, SumOfSqs = round(SumOfSqs, 3),
              R2 = round(R2, 3), `F` = round(`F`, 3), Pvalue = round(`Pr..F.`,3),
              Omega2_partial = round(Omega2_partial,3)) %>% column_to_rownames(var = "Name")    
  g <- ggplot(wdat, aes(x = PC1, y = PC2, color = alpha_grp1)) + 
      geom_point(size = 2) + 
      scale_color_manual(values = alpha_grp1_color, name = "Faith PD") + 
      labs(x = x_lab, y = y_lab, title = "") +
      theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin)     
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_Global_by_alpha_grp1.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)
  
  g <- ggplot(wdat_center, aes(x = PC1, y = PC2, color = alpha_grp1)) + 
    geom_point(data = wdat, aes(x = PC1, y = PC2, color = alpha_grp1), size = 1, alpha = 0.7) + 
    geom_segment(data = wdat, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4, shape = 21, color = "black", aes(fill = alpha_grp1)) + 
    scale_color_manual(values = alpha_grp1_color, 
                       name = "Faith PD", aesthetics = c("color","fill")) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin)  
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_Global_by_alpha_grp1_center.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)  
  ## alpha grp 2
  wrda_fit <- envfit(wrda ~ alpha_grp2, data = wmeta)
  wdat_center2 <- wrda_fit$factors$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "alpha_grp2") %>%
    transmute(alpha_grp2 = str_remove(alpha_grp2,"alpha_grp2"),PC1 = MDS1, PC2 = MDS2)
  wdat2 <- wdat_pcoa %>% left_join(wdat_center2 %>% transmute(PC1.c = PC1, PC2.c = PC2, alpha_grp2), by = "alpha_grp2")
  g_tab <- read.csv(paste0(temp_out,"adonis_test_", dd, "_Global_alpha2_only.csv")) %>% 
    filter(!is.na(Omega2_partial)) %>% 
    transmute(Name = Variable, Df = Df, SumOfSqs = round(SumOfSqs, 3),
              R2 = round(R2, 3), `F` = round(`F`, 3), Pvalue = round(`Pr..F.`,3),
              Omega2_partial = round(Omega2_partial,3)) %>% column_to_rownames(var = "Name") 
    
  g <- ggplot(wdat2, aes(x = PC1, y = PC2, color = alpha_grp2)) + 
      geom_point(size = 2) + 
      scale_color_manual(values = alpha_grp2_color, name = "Faith PD") + 
      labs(x = x_lab, y = y_lab, title = "") +
      theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin) 
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_Global_by_alpha_grp2.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)  
  
  g <- ggplot(wdat_center2, aes(x = PC1, y = PC2, color = alpha_grp2)) + 
    geom_point(data = wdat2, aes(x = PC1, y = PC2, color = alpha_grp2), size = 1, alpha = 0.7) + 
    geom_segment(data = wdat2, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4, shape = 21, color = "black", aes(fill = alpha_grp2)) + 
    scale_color_manual(values = alpha_grp2_color, 
                       name = "Faith PD", aesthetics = c("color","fill")) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold')) 
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin) 
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_Global_by_alpha_grp2_center.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)  
}

```

## PCoA combine in global
```{r}
dat_alpha <- left_join(alpha_boot, meta_all, by = "SampleID")
alpha <- rbind(dat_alpha %>% filter(!is.na(Country), Project %in% c("Global", "Upside", "Milkbank")),
               dat_alpha %>% filter(!is.na(Country), Project=="PR") %>% filter(TimePoint == 1200))
#hist(alpha$Faith_PD)
alpha_cutoff <- 10
meta<- alpha %>% mutate(alpha_grp1 = ifelse(Faith_PD > alpha_cutoff, ">10", "<=10"),
                          alpha_grp2 = ntile(Faith_PD, 3)) %>% 
  mutate(alpha_grp2 = case_when(alpha_grp2 == 1 ~ "Low",
                                alpha_grp2 == 2 ~ "Medium",
                                alpha_grp2 == 3 ~ "High"))

for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta = meta %>% filter(SampleID %in% labels(wdist))
  wdist <- dist_subset(wdist, wmeta$SampleID) 
  wrda <- dbrda(wdist ~ 1, sqrt.dist = TRUE)
  wrda_eig <- eigenvals(wrda)
  varPC1 <- round(wrda_eig["MDS1"]/sum(wrda_eig)*100,2)
  varPC2 <- round(wrda_eig["MDS2"]/sum(wrda_eig)*100,2)
  wdat_pcoa <- merge(scores(wrda, display = "sites", choices = c(1,2)),
                    wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = MDS1, PC2 = MDS2)
  x_lab <- paste0("PC1 (", varPC1, "%)")
  y_lab <- paste0("PC2 (", varPC2, "%)")
  ## alpha
  wrda_fit <- envfit(wrda ~ alpha_grp2, data = wmeta)
  wdat_center2 <- wrda_fit$factors$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "alpha_grp2") %>%
    transmute(alpha_grp2 = str_remove(alpha_grp2,"alpha_grp2"),PC1 = MDS1, PC2 = MDS2)
  wdat2 <- wdat_pcoa %>% left_join(wdat_center2 %>% transmute(PC1.c = PC1, PC2.c = PC2, alpha_grp2), by = "alpha_grp2")
  g_alpha <- ggplot(wdat_center2, aes(x = PC1, y = PC2, color = alpha_grp2)) + 
    geom_point(data = wdat2, aes(x = PC1, y = PC2, color = alpha_grp2), size = 0.7, alpha = 0.7) + 
    geom_segment(data = wdat2, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 2, shape = 21, color = "black", aes(fill = alpha_grp2)) + 
    scale_color_manual(values = alpha_grp2_color, 
                       name = "Faith PD", aesthetics = c("color","fill")) + 
    labs(x = "", y = "", title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = 'transparent'),
                           plot.background = element_rect(fill = 'transparent', color = NA),
                           panel.grid = element_blank(),
                           axis.ticks = element_blank(),
                           axis.text = element_blank(),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'))    
  ## Country
  wrda_fit <- envfit(wrda ~ Country, data = wmeta)
  wdat_center <- wrda_fit$factors$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "Country") %>%
    transmute(Country = str_remove(Country,"Country"),PC1 = MDS1, PC2 = MDS2) %>%
    left_join(wmeta %>% dplyr::select(Country, Continent) %>% unique(),by = "Country")
  wdat <- wdat_pcoa %>% left_join(wdat_center %>% transmute(PC1.c = PC1, PC2.c = PC2, Country), by = "Country")
  g_country <- ggplot(wdat_center, aes(x = PC1, y = PC2, color = Continent)) + 
    geom_point(data = wdat, aes(x = PC1, y = PC2, color = Continent), size = 1, alpha = 0.5) + 
    geom_segment(data = wdat, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4,shape = 21, color = "black", aes(fill = Continent)) + 
    geom_text_repel(aes(label = Country), color = "black") + 
    scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))    
  ## Continent center
  wrda_fit <- envfit(wrda ~ Continent, data = wmeta)
  wdat_center2 <- wrda_fit$factors$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "Continent") %>%
    transmute(Continent = str_remove(Continent,"Continent"),PC1 = MDS1, PC2 = MDS2)
  wdat2 <- wdat_pcoa %>% left_join(wdat_center2 %>% transmute(PC1.c = PC1, PC2.c = PC2, Continent), by = "Continent")
  g_continent <- ggplot(wdat_center2, aes(x = PC1, y = PC2, color = Continent)) + 
    geom_point(data = wdat2, aes(x = PC1, y = PC2, color = Continent), size = 1, alpha = 0.5) + 
    geom_segment(data = wdat2, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4, shape = 21, color = "black", aes(fill = Continent)) + 
    geom_text_repel(aes(label = Continent), color = "black") + 
    scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold')) 
  
  ## 
  if (dd == "uu"){
    g <- g_continent + 
      inset_element(g_alpha, left = 0.6, bottom = 0.6, right = 1, top = 1, align_to = 'panel') + 
      plot_layout(guides = 'collect')
  }else if(dd == "wu"){
    g <- g_continent + 
      inset_element(g_alpha, left = 0, bottom = 0, right = 0.4, top = 0.4, align_to = 'panel') + 
      plot_layout(guides = 'collect')
  }else if(dd == "bray"){
    g <- g_continent + 
      inset_element(g_alpha, left = 0, bottom = 0, right = 0.4, top = 0.4, align_to = 'panel') + 
      plot_layout(guides = 'collect')
  }else if(dd == "jaccard"){
    g <- g_continent + 
      inset_element(g_alpha, left = 0, bottom = 0, right = 0.4, top = 0.4, align_to = 'panel') + 
      plot_layout(guides = 'collect')
  }
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_Global_continent_alpha2.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
}

```

## different way for PcoA
```{r}
dd = "bray"
wdist <- get(paste0("DM_",dd))
wmeta = meta_all %>% filter(SampleID %in% labels(wdist), !is.na(Country))
wdist <- dist_subset(wdist, wmeta$SampleID)  

## dbRDA
wrda <- dbrda(wdist ~ 1, sqrt.dist = TRUE)
wrda_eig <- eigenvals(wrda)
varPC1_1 <- round(wrda_eig["MDS1"]/sum(wrda_eig)*100,2)
varPC2_1 <- round(wrda_eig["MDS2"]/sum(wrda_eig)*100,2)
x_lab_1 <- paste0("PC1 (", varPC1_1, "%)")
y_lab_1 <- paste0("PC2 (", varPC2_1, "%)")
wdat_wrda <- merge(scores(wrda, display = "sites", choices = c(1,2)),
                   wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
  rename(PC1 = MDS1, PC2 = MDS2)
wrda_fit <- envfit(wrda ~ Country, data = wmeta)
wdat_center_1 <- wrda_fit$factors$centroids %>% as.data.frame() %>% 
  rownames_to_column(var = "Country") %>%
  transmute(Country = str_remove(Country,"Country"),PC1 = MDS1, PC2 = MDS2) %>%
  left_join(wmeta %>% dplyr::select(Country, Continent) %>% unique(),by = "Country")
wdat_1 <- wdat_wrda %>% left_join(wdat_center_1 %>% transmute(PC1.c = PC1, PC2.c = PC2, Country), by = "Country")

g1 <- ggplot(wdat_center_1, aes(x = PC1, y = PC2, color = Continent)) + 
  geom_point(data = wdat_1, aes(x = PC1, y = PC2, color = Continent), size = 1, alpha = 0.5) + 
  geom_segment(data = wdat_1, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
  geom_point(size = 4,shape = 21, color = "black", aes(fill = Continent)) + 
  geom_text_repel(aes(label = Country), color = "black") + 
  scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
  labs(x = x_lab_1, y = y_lab_1, title = "vegan::dbrda") 

## wcmdscale
wpcoa_v <- wcmdscale(d = wdist, eig = TRUE)
varPC1_2 <- round(wpcoa_v$eig[1]/sum(wpcoa_v$eig)*100,2)
varPC2_2 <- round(wpcoa_v$eig[2]/sum(wpcoa_v$eig)*100,2)
x_lab_2 <- paste0("PC1 (", varPC1_2, "%)")
y_lab_2 <- paste0("PC2 (", varPC2_2, "%)")
wdat_pcoa_v <- merge(wpcoa_v$points[,c(1,2)],
                   wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
  rename(PC1 = Dim1, PC2 = Dim2)
wpcoa_fit <- envfit(wpcoa_v ~ Country, data = wmeta)
wdat_center_2 <- wpcoa_fit$factors$centroids %>% as.data.frame() %>% 
  rownames_to_column(var = "Country") %>%
  transmute(Country = str_remove(Country,"Country"),PC1 = Dim1, PC2 = Dim2) %>%
  left_join(wmeta %>% dplyr::select(Country, Continent) %>% unique(),by = "Country")
wdat_2 <- wdat_pcoa_v %>% left_join(wdat_center_2 %>% transmute(PC1.c = PC1, PC2.c = PC2, Country), by = "Country")

g2 <- ggplot(wdat_center_2, aes(x = PC1, y = PC2, color = Continent)) + 
  geom_point(data = wdat_2, aes(x = PC1, y = PC2, color = Continent), size = 1, alpha = 0.5) + 
  geom_segment(data = wdat_2, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
  geom_point(size = 4,shape = 21, color = "black", aes(fill = Continent)) + 
  geom_text_repel(aes(label = Country), color = "black") + 
  scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
  labs(x = x_lab_2, y = y_lab_2, title = "vegan::wcmdscale") 

## pcoa
pcoa_3 <- pcoa(wdist)
varPC1_3 <- round(pcoa_3$values$Relative_eig[1]*100, 2)
varPC2_3 <- round(pcoa_3$values$Relative_eig[2]*100, 2)
x_lab_3 <- paste0("PC1 (", varPC1_3, "%)")
y_lab_3 <- paste0("PC2 (", varPC2_3, "%)")
pcoa_dat_3 <- merge(pcoa_3$vectors[,1:2], 
                    wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
  rename(PC1 = Axis.1, PC2 = Axis.2)
wdat_center_3 <- pcoa_dat_3 %>% group_by(Country) %>% 
        summarise(across(starts_with("PC"), mean)) %>% dplyr::select(-PCRConc) %>% 
  left_join(wmeta %>% dplyr::select(Country, Continent) %>% unique(),by = "Country")
wdat_3 <- pcoa_dat_3 %>% left_join(wdat_center_3%>% transmute(PC1.c = PC1, PC2.c = PC2, Country), by = "Country")
g3 <- ggplot(wdat_center_3, aes(x = PC1, y = PC2, color = Continent)) + 
  geom_point(data = wdat_3, aes(x = PC1, y = PC2, color = Continent), size = 1, alpha = 0.5) + 
  geom_segment(data = wdat_3, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
  geom_point(size = 4,shape = 21, color = "black", aes(fill = Continent)) + 
  geom_text_repel(aes(label = Country), color = "black") + 
  scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
  labs(x = x_lab_3, y = y_lab_3, title = "ape::pcoa") 

library(patchwork)
g <- g1+g2+g3
```
