==> code/06_new_global.Rmd <==
---
title: "Alpha diversity"
author: "Haipeng Sun"
date: "8/1/2022"
output: html_document
---

## library
```{r setup, message=FALSE, warning=FALSE}
library(phyloseq)
library(purrr)
library(furrr)
library(picante)
library(vegan)
library(MASS)
library(ggpubr)
library(effectsize)
library(ggtext)
library(glue)
library(tidyverse)
library(patchwork)
options(scipen=999)
```

## prepare (run first time only)
### load data
```{r}
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/phylo_meta.Rdata")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_uu.RData")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_wu.RData")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_bray.RData")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_jaccard.RData")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/alpha.Rdata")
```

### function
```{r}
comp_kw_aov = function(wdat, x, y){
  #input data, respond variable y, categorize variable x
  require(agricolae)
  ktest1 = kruskal(wdat[y], wdat[x], alpha = 0.05, p.adj="fdr", group=T, main = "wdat", console=FALSE)
  out_g_1 = ktest1$groups %>% rownames_to_column(var = "Var") %>% dplyr::select(Var, groups) %>% rename(KW_group = groups)
  ktest2 = kruskal(wdat[y], wdat[x], alpha = 0.05, p.adj="none", group=F, main = "wdat", console=FALSE)
  out_c_1 = ktest2$comparison %>% rownames_to_column(var = "comp") %>% dplyr::select(comp, pvalue) %>% rename(KW_pvalue = pvalue)
  fit = aov(as.formula(paste0(y, "~", x)), wdat)
  tukey1 = HSD.test(fit, x, group = T)
  out_g_2 = tukey1$groups %>% rownames_to_column(var = "Var") %>% dplyr::select(Var, groups) %>% rename(aov_group = groups)
  tukey2 = HSD.test(fit, x, group = F)
  out_c_2 = tukey2$comparison %>% rownames_to_column(var = "comp") %>% dplyr::select(comp, pvalue) %>% rename(aov_pvalue = pvalue)
  out_g = merge(out_g_1, out_g_2,by = 1)
  out_c = merge(out_c_1, out_c_2,by = 1)
  res = list(out_g, out_c)
  names(res) = c("grp","comp")
  res
}

my_summarize = function(x){
  data.frame(
    N = n(),
    N.nna = sum(!is.na(x)),
    Min = min(x,na.rm = T),
    Med = median(x, na.rm = T),
    Mean = mean(x, na.rm = T),
    Max = max(x,na.rm = T),
    SD = sd(x,na.rm = T)
  )
}

```

### global meta
```{r}
meta_common <- read.csv("~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/fungi_sample_common.csv") %>% 
  select(-X)
meta_common <- meta_common %>% filter(Project != "Covid Mom")

meta_global <- rbind(meta_all %>% filter(!is.na(Country), Project %in% c("Global", "Milkbank")),
                     meta_all %>% filter(!is.na(Country), Project=="PR") %>% 
                       filter(TimePoint == 1200),
                     meta_all %>% filter(!is.na(Country), Project=="Family") %>% 
                       filter(TimePoint == 1200)
) 

meta_global <- meta_global[!duplicated(meta_global$Mother_ID),] %>% 
  mutate(SampleID_clean = str_remove_all(SampleID, " "))

meta_global <- rbind(meta_global %>% filter(!Mother_ID %in% meta_common$Mother_ID),
                     meta_common %>% filter(!duplicated(Mother_ID)))
meta_global <- meta_global %>% filter(Project != "Covid Mom")

meta_global <- meta_global %>% left_join(sample_depth %>% dplyr::select(SampleID, decontam_cleaned), by = "SampleID")

```

### global sample filter
```{r}
phylo_global <- prune_samples(sample_names(phylo_all) %in% meta_global$SampleID, 
                              x = phylo_all)

contam_asv <- read_csv("~/Dropbox/Rutgers/20220606_RUN41_milk/Data/output/decontam_details.csv",show_col_types = FALSE)
contam_id <- contam_asv %>% filter(prev == "Y") %>% pull(id)
phylo_global_decontam <- prune_taxa(!(taxa_names(phylo_global) %in% contam_id), phylo_global)

remove_id <- c(taxa %>% filter(str_detect(Kingdom, "Unassigned")) %>% pull(id),
               taxa %>% filter(Phylum == "p_") %>% pull(id),
               taxa %>% filter(str_detect(Kingdom, "Eukaryota")) %>% pull(id),
               taxa %>% filter(str_detect(taxa_long, "Chloroplast")) %>% pull(id),
               taxa %>% filter(str_detect(taxa_long, "Mitochondria")) %>% pull(id)
) %>% unique()
phylo_global_decontam_clean <- prune_taxa(!(taxa_names(phylo_global_decontam) %in% remove_id), phylo_global_decontam)

mean(sample_sums(phylo_global))
mean(sample_sums(phylo_global_decontam))
mean(sample_sums(phylo_global_decontam_clean))

sample_depth_global <- cbind(sample_sums(phylo_global),
                             sample_sums(phylo_global_decontam),
                             sample_sums(phylo_global_decontam_clean)) %>% 
  as.data.frame() %>% 
  rename(original=V1, decontamed=V2, decontam_cleaned=V3) %>%
  merge(., meta_all, by.x=0, by.y="SampleID") %>% 
  transmute(SampleID=Row.names, DNA_Conc_ng_uL, Country, Continent,
            original, decontamed, decontam_cleaned) %>% 
  arrange(decontam_cleaned)

ft <- otu_table(phylo_global_decontam_clean) %>% as.matrix() %>% as.data.frame() %>% t()

coverage_stat <- ft %>% as.data.frame() %>% 
  rownames_to_column(var = "SampleID") %>% 
  pivot_longer(cols = -SampleID, names_to = "id", values_to = "count") %>% 
  filter(count > 0) %>% group_by(SampleID) %>%
  summarise(n_seqs = sum(count),
            n_singles = sum(count==1),
            goods = 100*(1 - n_singles/n_seqs)) %>% 
  ungroup()

rarecurve_data = rarecurve(ft, step = 50, tidy = T)

coverage_stat %>% ggplot(aes(x=n_seqs)) + 
  geom_histogram() + 
  labs(x="Depth", title = "Histograph of sequencing depth") 

coverage_stat %>% arrange(n_seqs) %>%
  ggplot(aes(x=1:nrow(.), y=n_seqs))+ 
  geom_point() + 
  labs(x = "", y="Depth", title = "Sequencing depth low to high")

coverage_stat %>%ggplot(aes(x=n_seqs, y=goods)) + 
  geom_point() + 
  geom_hline(yintercept = 99, color = "red") + 
  labs(x = "Depth", y = "Goods coverage", title = "Goods coverage")

rarecurve_data %>% ggplot(aes(x=Sample, y=Species, group=Site)) + 
  geom_line(alpha = 0.2) + 
  scale_y_log10() +
  labs(x = "Depth", y = "Number of ASVs") + 
  coord_cartesian(xlim = c(0,1000)) 

### set depth to 2857
depth = 2857
g <- coverage_stat %>%ggplot(aes(x=n_seqs, y=goods)) + 
  geom_point() + 
  geom_hline(yintercept = 99, color = "red") + 
  geom_label(data = data.frame(n_seqs = c(60000,12000), 
                               goods = c(98, 80), 
                               label = c("99 Goods coverage", paste0("Rarefy depth = ", depth))), 
             aes(label = label)) + 
  geom_vline(xintercept = depth, color = "black") + 
  labs(x = "Depth", y = "Goods coverage", title = "Goods coverage")
ggsave(filename = "~/Dropbox/Rutgers/20220606_RUN41_milk/Data/output/global/Goods_coverage_depth.pdf", 
       device = cairo_pdf, plot = g, width = 7, height = 5, units = "in", dpi = 300)

g <- ggplot(data = rarecurve_data %>% filter(Site %in% sample_depth_global$SampleID), 
            aes(x=Sample, y=Species, group=Site)) + 
  geom_line(alpha = 0.3) + 
  geom_line(data = rarecurve_data %>% filter(!(Site %in% sample_depth_global$SampleID)),
            alpha = 1, color = "red") +
  geom_vline(xintercept = depth, color = "black") + 
  geom_label(data = data.frame(Sample = c(2600), 
                               Species = c(900), 
                               Site = "",
                               label = paste0("Rarefy depth = ", depth)), 
             aes(label = label)) + 
  scale_y_log10() +
  labs(x = "Depth", y = "Number of ASVs") + 
  coord_cartesian(xlim = c(0,15000)) 
ggsave(filename = "~/Dropbox/Rutgers/20220606_RUN41_milk/Data/output/global/depth_rarefy.pdf", 
       device = cairo_pdf, plot = g, width = 7, height = 5, units = "in", dpi = 300)
```


### global 16S data
```{r}
phylo_global_decontam_clean_rf = rarefy_even_depth(phylo_global_decontam_clean, 
                                                   sample.size = depth, 
                                                   rngseed = 3, 
                                                   replace = F)

# alpha
a1 <- picante::pd(t(phylo_global_decontam_clean_rf@otu_table@.Data), 
                  phylo_global_decontam_clean_rf@phy_tree, include.root = F)
a2 <- estimate_richness(phylo_global_decontam_clean_rf, measures="Shannon")
alpha_global <- merge(a1, a2, by=0)
names(alpha_global) <- c("SampleID","Faith_PD", "Observed_ASVs","Shannon_Index")
alpha_global$Pielou_evenness <- alpha_global$Shannon_Index/log(alpha_global$Observed_ASVs)

meta_global_rf <- meta_global %>% filter(SampleID %in% alpha_global$SampleID)

# beta
DM_uu_global <- UniFrac(phylo_global_decontam_clean_rf, weighted = FALSE)  
DM_wu_global <- UniFrac(phylo_global_decontam_clean_rf, weighted = TRUE)
ft <- t(phylo_global_decontam_clean_rf@otu_table@.Data)
DM_bray_global <- vegdist(ft,method = "bray")
DM_jaccard_global <- vegdist(ft,method = "jaccard")

## manually correct phylum name
# Firmicutes ~ Bacillota
# Proteobacteria ~ Pseudomonadota
# Actinobacteriota ~ Actinomycetota
# Bacteroidetes ~ Bacteroidota
# Campilobacterota ~ Campylobacterota
# Spirochaetota
# Deinococcota 
# Verrucomicrobiota
taxa$Phylum %>% unique() %>% sort()
taxa <- taxa %>% 
  mutate(Phylum = case_when(str_detect(Phylum, "Firmicutes") ~ "p_Bacillota",
                            str_detect(Phylum, "Proteobacteria") ~ "p_Pseudomonadota",
                            str_detect(Phylum, "Actinobacteriota") ~ "p_Actinomycetota",
                            str_detect(Phylum, "Bacteroidetes") ~ "p_Bacteroidota",
                            str_detect(Phylum, "Campilobacterota") ~ "p_Campylobacterota",
                            TRUE ~ Phylum))
taxa$Phylum %>% unique() %>% sort()

taxa <- taxa %>% 
  mutate(Class_l = case_when(Class == "c_" ~ str_c(Phylum, Class, sep = ";"),
                             TRUE ~ Class),
         Order_l = case_when(Order == "o_" ~ str_c(Class_l, Order, sep = ";"),
                             str_detect(Order, "uncultured") ~ str_c(Class_l, Order, sep = ";"),
                             TRUE ~ Order),
         Family_l = case_when(Family == "f_" ~ str_c(Order_l, Family, sep = ";"),
                              str_detect(Family, "uncultured") ~ str_c(Order_l, Family, sep = ";"),
                              TRUE ~ Family),
         Genus_l = case_when(Genus == "g_" ~ str_c(Family_l, Genus, sep = ";"),
                             str_detect(Genus, "uncultured") ~ str_c(Family_l, Genus, sep = ";"),
                             TRUE ~ Genus),
         Species_l = case_when(Species == "s_" ~ str_c(Genus_l, Species, sep = ";"),
                               str_detect(Species, "uncultured|metagenome") ~ str_c(Genus_l, Species, sep = ";"),
                               TRUE ~ Species))

## summary data flow
sink("~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/output/global/sample_project_sum.txt")
table(meta_all$Project)
table(meta_global$Project)
table(meta_global_rf$Project)
print("fungi")
meta_common %>% filter(!duplicated(Mother_ID)) %>% pull(Project) %>% table()
sink()

```

### save
```{r}
save(Continent_color, Country_color, Country_font_color, Country_labels,
     alpha_global, DM_bray_global, DM_jaccard_global, DM_uu_global, DM_wu_global, 
     depth, meta_global, meta_global_rf, taxa, 
     phylo_global, phylo_global_decontam_clean, phylo_global_decontam_clean_rf,
     comp_kw_aov, my_summarize, rarecurve_data, sample_depth_global, 
     file = "~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/Global.Rdata")
```


## load global data
```{r}
load("~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/Global.Rdata")
temp_out = "~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/output/global/"

```


### global summary
```{r}
meta_global_rf <- meta_global_rf %>% 
  mutate(Maternal_BMI = Maternal_weight_kg/(Maternal_height_cm/100)/(Maternal_height_cm/100))

meta_global_rf %>% count(Country) %>% as.data.frame()
meta_global_rf %>% count(Continent) %>% print()
meta_global_rf %>% count(Birth_mode) %>% print()

meta_global_rf %>% summarise(my_summarize(Maternal_age)) %>% print()
meta_global_rf %>% summarise(my_summarize(Maternal_weight_kg)) %>% print()
meta_global_rf %>% summarise(my_summarize(Maternal_height_cm)) %>% print()
meta_global_rf %>% summarise(my_summarize(Maternal_BMI)) %>% print()
meta_global_rf %>% summarise(my_summarize(Time_post_partum_day)) %>% print()

a <- meta_global_rf %>% count(Continent, Country)
write.csv(a, file = paste0(temp_out,"Global_samples_by_Country.csv"),row.names = FALSE, quote = FALSE)

my_summarize2 = function(x){
  data.frame(
    N = n(),
    #N.nna = sum(!is.na(x)),
    Min = min(x,na.rm = T),
    Q25 = quantile(x, 0.25),
    Med = median(x, na.rm = T),
    Mean = mean(x, na.rm = T),
    Q75 = quantile(x, 0.75),
    Max = max(x,na.rm = T),
    Sum = sum(x, na.rm = T)
  )
}

meta_global %>% summarise(my_summarize2(Seq_count)) %>% print()
meta_global %>% summarise(my_summarize2(decontam_cleaned)) %>% print()


sum(!is.na(meta_global_rf$Maternal_age))
sum(!is.na(meta_global_rf$Maternal_weight_kg))
sum(!is.na(meta_global_rf$Maternal_height_cm))
sum(!is.na(meta_global_rf$Birth_mode))
sum(!is.na(meta_global_rf$Time_post_partum_day))

my_summarize3 = function(x){
  data.frame(
    N = n(),
    N.nna = sum(!is.na(x)),
    Med = median(x, na.rm = T),
    Mean = mean(x, na.rm = T),
    SD = sd(x, na.rm = T)
  )
}

meta_global %>% 
  #group_by(Country) %>% 
  summarise(my_summarize3(Maternal_age)) %>% 
  #write_csv(.,file = paste0(temp_out,"Global_samples_Maternal_age.csv"))

meta_global %>% group_by(Country) %>% 
  summarise(my_summarize3(Maternal_weight_kg)) %>% 
  write_csv(.,file = paste0(temp_out,"Global_samples_Maternal_weight_kg.csv"))

meta_global %>% group_by(Country) %>% 
  summarise(my_summarize3(Time_post_partum_day)) %>% 
  write_csv(.,file = paste0(temp_out,"Global_samples_Time_post_partum_day.csv"))

meta_global %>% group_by(Country) %>% 
  count(Birth_mode) %>% pivot_wider(id_cols = Country, names_from = Birth_mode, values_from = n) %>% 
  write_csv(.,file = paste0(temp_out,"Global_samples_Birth_mode.csv"))


```


## alpha
### function and data
```{r}
load("~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/Global.Rdata")

comp_kw_aov = function(wdat, x, y){
  #input data, respond variable y, categorize variable x
  require(agricolae)
  ktest1 = kruskal(wdat[y], wdat[x], alpha = 0.05, p.adj="fdr", group=T, main = "wdat", console=FALSE)
  out_g_1 = ktest1$groups %>% rownames_to_column(var = "Var") %>% dplyr::select(Var, groups) %>% rename(KW_group = groups)
  ktest2 = kruskal(wdat[y], wdat[x], alpha = 0.05, p.adj="fdr", group=F, main = "wdat", console=FALSE)
  out_c_1 = ktest2$comparison %>% rownames_to_column(var = "comp") %>% dplyr::select(comp, pvalue) %>% rename(KW_pvalue = pvalue)
  fit = aov(as.formula(paste0(y, "~", x)), wdat)
  tukey1 = HSD.test(fit, x, group = T)
  out_g_2 = tukey1$groups %>% rownames_to_column(var = "Var") %>% dplyr::select(Var, groups) %>% rename(aov_group = groups)
  tukey2 = HSD.test(fit, x, group = F)
  out_c_2 = tukey2$comparison %>% rownames_to_column(var = "comp") %>% dplyr::select(comp, pvalue) %>% rename(aov_pvalue = pvalue)
  out_g = merge(out_g_1, out_g_2,by = 1)
  out_c = merge(out_c_1, out_c_2,by = 1)
  res = list(out_g, out_c)
  names(res) = c("grp","comp")
  res
}
```


### alpha linear model global
```{r}
temp_out <- "~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/output/global/alpha/"
metric <- names(alpha_global)[2:5]
names(metric) <- names(alpha_global)[2:5]
dat <- alpha_global %>% left_join(meta_global, by = "SampleID")

# country
alpha_lm_coef <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Country")), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_coef$Metrics = a
  res_coef
})
alpha_lm_coef$padj = round(p.adjust(alpha_lm_coef$`Pr(>|t|)`), digits = 4)
alpha_lm_coef$pvalue = round(alpha_lm_coef$`Pr(>|t|)`, digits = 4)
write.csv(alpha_lm_coef, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_coef_Country.csv"))

alpha_lm_omega <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Country")), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_anova$Metrics = a
  res_anova
})
alpha_lm_omega$padj = round(p.adjust(alpha_lm_omega$`Pr(>F)`), digits = 4)
alpha_lm_omega$pvalue = round(alpha_lm_omega$`Pr(>F)`, digits = 4)
write.csv(alpha_lm_omega, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_omega_Country.csv"))

# Continent
alpha_lm_coef <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Continent")), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_coef$Metrics = a
  res_coef
})
alpha_lm_coef$padj = round(p.adjust(alpha_lm_coef$`Pr(>|t|)`), digits = 4)
alpha_lm_coef$pvalue = round(alpha_lm_coef$`Pr(>|t|)`, digits = 4)
write.csv(alpha_lm_coef, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_coef_Continent.csv"))

alpha_lm_omega <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Continent")), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_anova$Metrics = a
  res_anova
})
alpha_lm_omega$padj = round(p.adjust(alpha_lm_omega$`Pr(>F)`), digits = 4)
alpha_lm_omega$pvalue = round(alpha_lm_omega$`Pr(>F)`, digits = 4)
write.csv(alpha_lm_omega, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_omega_Continent.csv"))
## Time_post_partum_day
alpha_lm_coef <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Time_post_partum_day" )), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_coef$Metrics = a
  res_coef
})
alpha_lm_coef$padj = round(p.adjust(alpha_lm_coef$`Pr(>|t|)`), digits = 4)
alpha_lm_coef$pvalue = round(alpha_lm_coef$`Pr(>|t|)`, digits = 4)
write.csv(alpha_lm_coef, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_coef_partum_day.csv"))

alpha_lm_omega <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Time_post_partum_day")), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_anova$Metrics = a
  res_anova
})
alpha_lm_omega$padj = round(p.adjust(alpha_lm_omega$`Pr(>F)`), digits = 4)
alpha_lm_omega$pvalue = round(alpha_lm_omega$`Pr(>F)`, digits = 4)
write.csv(alpha_lm_omega, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_omega_partum_day.csv"))

## Maternal_age
alpha_lm_coef <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Maternal_age" )), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_coef$Metrics = a
  res_coef
})
alpha_lm_coef$padj = round(p.adjust(alpha_lm_coef$`Pr(>|t|)`), digits = 4)
alpha_lm_coef$pvalue = round(alpha_lm_coef$`Pr(>|t|)`, digits = 4)
write.csv(alpha_lm_coef, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_coef_Maternal_age.csv"))

alpha_lm_omega <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Maternal_age")), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_anova$Metrics = a
  res_anova
})
alpha_lm_omega$padj = round(p.adjust(alpha_lm_omega$`Pr(>F)`), digits = 4)
alpha_lm_omega$pvalue = round(alpha_lm_omega$`Pr(>F)`, digits = 4)
write.csv(alpha_lm_omega, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_omega_Maternal_age.csv"))

## Birthmode
alpha_lm_coef <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Birth_mode" )), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_coef$Metrics = a
  res_coef
})
alpha_lm_coef$padj = round(p.adjust(alpha_lm_coef$`Pr(>|t|)`), digits = 4)
alpha_lm_coef$pvalue = round(alpha_lm_coef$`Pr(>|t|)`, digits = 4)
write.csv(alpha_lm_coef, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_coef_Birth_mode.csv"))

alpha_lm_omega <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Birth_mode")), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_anova$Metrics = a
  res_anova
})
alpha_lm_omega$padj = round(p.adjust(alpha_lm_omega$`Pr(>F)`), digits = 4)
alpha_lm_omega$pvalue = round(alpha_lm_omega$`Pr(>F)`, digits = 4)
write.csv(alpha_lm_omega, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_omega_Birth_mode.csv"))

## Maternal_weight_kg
alpha_lm_coef <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Maternal_weight_kg" )), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_coef$Metrics = a
  res_coef
})
alpha_lm_coef$padj = round(p.adjust(alpha_lm_coef$`Pr(>|t|)`), digits = 4)
alpha_lm_coef$pvalue = round(alpha_lm_coef$`Pr(>|t|)`, digits = 4)
write.csv(alpha_lm_coef, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_coef_Maternal_weight_kg.csv"))

alpha_lm_omega <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Maternal_weight_kg")), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_anova$Metrics = a
  res_anova
})
alpha_lm_omega$padj = round(p.adjust(alpha_lm_omega$`Pr(>F)`), digits = 4)
alpha_lm_omega$pvalue = round(alpha_lm_omega$`Pr(>F)`, digits = 4)
write.csv(alpha_lm_omega, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_omega_Maternal_weight_kg.csv"))

## All
alpha_lm_coef <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Birth_mode + Maternal_weight_kg + Maternal_age +  Time_post_partum_day + Country" )), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_coef$Metrics = a
  res_coef
})
alpha_lm_coef$padj = round(p.adjust(alpha_lm_coef$`Pr(>|t|)`), digits = 4)
alpha_lm_coef$pvalue = round(alpha_lm_coef$`Pr(>|t|)`, digits = 4)
write.csv(alpha_lm_coef, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_coef_All.csv"))

alpha_lm_omega <- map_dfr(metric, function(a){
  fit <- lm(as.formula(paste0(a, "  ~ Birth_mode + Maternal_weight_kg + Maternal_age +  Time_post_partum_day + Country")), data = dat)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var = "Var") %>% filter(Var != "(Intercept)")
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter")
  res_anova$Metrics = a
  res_anova
})
alpha_lm_omega$padj = round(p.adjust(alpha_lm_omega$`Pr(>F)`), digits = 4)
alpha_lm_omega$pvalue = round(alpha_lm_omega$`Pr(>F)`, digits = 4)
write.csv(alpha_lm_omega, row.names = FALSE,
          file = paste0(temp_out,"alpha_lm_omega_All.csv"))

``` 


### alpha in Global project plot
```{r}
temp_out <- "~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/output/global/alpha/"
metric <- names(alpha_global)[2:5]
names(metric) <- names(alpha_global)[2:5]
dat <- alpha_global %>% left_join(meta_global, by = "SampleID")

# Country
alpha_Country_grp <- map_dfr(metric, function(a){
  res <- comp_kw_aov(wdat = dat, x = "Country", y = a)
  res$grp$Metrics = a
  res$grp
})
write.csv(alpha_Country_grp, row.names = FALSE,
          file = paste0(temp_out,"alpha_by_Country_kw_aov_grp.csv"))

alpha_Country_pvalue <- map_dfr(metric, function(a){
  res <- comp_kw_aov(wdat = dat, x = "Country", y = a)
  res$comp$Metrics = a
  res$comp
})
alpha_Country_pvalue$KW_pvalue_adj <- p.adjust(alpha_Country_pvalue$KW_pvalue)
alpha_Country_pvalue$aov_pvalue_adj <- p.adjust(alpha_Country_pvalue$aov_pvalue)
write.csv(alpha_Country_pvalue, row.names = FALSE, 
          file = paste0(temp_out,"alpha_by_Country_kw_pvalue.csv"))

alpha_Country_summary <- dat %>% dplyr::select(Faith_PD:Pielou_evenness, Country) %>% 
  group_by(Country) %>% 
  summarise_all(my_summarize) %>% 
  ungroup() %>%
  pivot_longer(-Country,
               names_to = "Metrics",
               values_to = "value") %>% 
  unnest(cols = value) %>% 
  left_join(alpha_Country_grp, by = c("Country"="Var", "Metrics"="Metrics"))

wdat <- dat %>% 
  mutate(Country = factor(Country,
                          levels = names(Country_color),
                          ordered = T)) %>% 
  dplyr::select(SampleID, Country, Faith_PD:Pielou_evenness) %>% 
  pivot_longer(Faith_PD:Pielou_evenness, names_to = "Metrics", values_to = "value")

## pvalue
stat.test <- wdat %>% group_by(Metrics) %>% 
  rstatix::wilcox_test(value ~ Country) %>% 
  rstatix::adjust_pvalue(method = "fdr") %>%
  rstatix::add_significance("p.adj") %>% 
  rstatix::add_xy_position(x = "Country", dodge = 0.8) %>% 
  filter(p.adj.signif != "ns")

g <- ggplot(wdat, aes(Country, value)) + 
  geom_boxplot(aes(color = Country, fill = Country),alpha = 0.7,  outlier.alpha = 0) + 
  geom_point(aes(color = Country), position = position_jitterdodge()) + 
  geom_text(data = alpha_Country_summary, 
            aes(x =  Country, y = 1.08 * Max, label = KW_group)) + 
  scale_color_manual(values = Country_color, aesthetics = c("color","fill")) + 
  scale_x_discrete(labels=Country_labels) + 
  labs(x = "", y = "") +
  facet_wrap(~Metrics, nrow = 2, scales = "free") +
  theme_bw() + theme(#aspect.ratio = 0.8,
    legend.position="none",
    panel.background = element_rect(fill = NA),
    strip.text = element_text(size=12,color="black",face='bold'),
    axis.title = element_text(size=12,color="black",face='bold'),
    axis.text = element_text(size=12,color="black",face='bold'),
    axis.text.x = element_markdown(size=12, face='bold',angle=90,vjust = 0.5, hjust = 1),
    axis.text.y = element_text(size=12,color="black",face='bold'),
    legend.text = element_text(size=12,color="black",face='bold'),
    legend.title = element_text(size=12,color="black",face='bold'),
    title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out, "alpha_by_Country_full_letter.pdf"), device = cairo_pdf, 
       plot = g, width = 14, height = 10, units = "in", dpi = 300) 

g <- ggplot(wdat, aes(Country, value)) + 
  geom_boxplot(aes(color = Country, fill = Country),alpha = 0.7,  outlier.alpha = 0) + 
  geom_point(aes(color = Country), position = position_jitterdodge()) + 
  stat_pvalue_manual(stat.test, label = "p.adj.signif", tip.length = 0) +
  scale_color_manual(values = Country_color, aesthetics = c("color","fill")) + 
  scale_x_discrete(labels=Country_labels) + 
  labs(x = "", y = "") +
  facet_wrap(~Metrics, nrow = 2, scales = "free") +
  theme_bw() + theme(#aspect.ratio = 0.8,
    legend.position="none",
    panel.background = element_rect(fill = NA),
    strip.text = element_text(size=12,color="black",face='bold'),
    axis.title = element_text(size=12,color="black",face='bold'),
    axis.text = element_text(size=12,color="black",face='bold'),
    axis.text.x = element_markdown(size=12, face='bold',angle=90,vjust = 0.5, hjust = 1),
    axis.text.y = element_text(size=12,color="black",face='bold'),
    legend.text = element_text(size=12,color="black",face='bold'),
    legend.title = element_text(size=12,color="black",face='bold'),
    title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out, "alpha_by_Country_full_pvalue.pdf"), device = cairo_pdf, 
       plot = g, width = 14, height = 10, units = "in", dpi = 300) 

# order by faith PD
Country_ord <- alpha_Country_summary %>% filter(Metrics == "Faith_PD") %>% 
  dplyr::select(Country, Med) %>% 
  left_join(meta_global %>% dplyr::select(Country, Continent) %>% unique(), by = "Country") %>% 
  arrange(Continent, desc(Med)) %>% pull(Country)
wdat$Country = factor(wdat$Country, levels = Country_ord, ordered = T)

g <- ggplot(wdat, aes(Country, value)) + 
  geom_boxplot(aes(color = Country, fill = Country),alpha = 0.7,  outlier.alpha = 0) + 
  geom_point(aes(color = Country), position = position_jitterdodge()) + 
  geom_text(data = alpha_Country_summary, 
            aes(x =  Country, y = 1.08 * Max, label = KW_group)) + 
  scale_color_manual(values = Country_color, aesthetics = c("color","fill")) + 
  scale_x_discrete(labels=Country_labels) + 
  labs(x = "", y = "") +
  facet_wrap(~Metrics, nrow = 2, scales = "free") +
  theme_bw() + theme(#aspect.ratio = 0.8,
    legend.position="none",
    panel.background = element_rect(fill = NA),
    strip.text = element_text(size=12,color="black",face='bold'),
    axis.title = element_text(size=12,color="black",face='bold'),
    axis.text = element_text(size=12,color="black",face='bold'),
    axis.text.x = element_markdown(size=12, face='bold',angle=90,vjust = 0.5, hjust = 1),
    axis.text.y = element_text(size=12,color="black",face='bold'),
    legend.text = element_text(size=12,color="black",face='bold'),
    legend.title = element_text(size=12,color="black",face='bold'),
    title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out, "alpha_by_Country_full_letter_sort_by_faith.pdf"), device = cairo_pdf, 
       plot = g, width = 14, height = 10, units = "in", dpi = 300) 


for (a in metric){
  wwdat <- wdat %>% filter(Metrics == a)
  wsummary <- alpha_Country_summary %>% filter(Metrics == a) %>%
    arrange(Med)
  ord = wsummary %>% pull(Country)
  wwdat$Country <- factor(wwdat$Country, levels = ord, ordered = T)
  g <- ggplot(wwdat, aes(Country, value)) + 
    geom_boxplot(aes(color = Country, fill = Country),alpha = 0.7,  outlier.alpha = 0) + 
    geom_point(aes(color = Country), position = position_jitterdodge()) + 
    geom_text(data = wsummary, 
              aes(x =  Country, y = 1.08 * Max, label = KW_group)) + 
    scale_color_manual(values = Country_color, aesthetics = c("color","fill")) + 
    scale_x_discrete(labels=Country_labels) + 
    labs(x = "", y = a) + 
    theme_bw() + theme(#aspect.ratio = 0.8,
      legend.position="none",
      panel.background = element_rect(fill = NA),
      strip.text = element_text(size=12,color="black",face='bold'),
      axis.title = element_text(size=12,color="black",face='bold'),
      axis.text = element_text(size=12,color="black",face='bold'),
      axis.text.x = element_markdown(size=12,face='bold',angle=90,vjust = 0.5, hjust = 1),
      axis.text.y = element_text(size=12,color="black",face='bold'),
      legend.text = element_text(size=12,color="black",face='bold'),
      legend.title = element_text(size=12,color="black",face='bold'),
      title = element_text(size=12,color="black",face='bold'))
  ggsave(filename = paste0(temp_out, "alpha_by_Country_single_", a, "_ordered.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300) 
}


# Continent
alpha_Continent_grp <- map_dfr(metric, function(a){
  res <- comp_kw_aov(wdat = dat, x = "Continent", y = a)
  res$grp$Metrics = a
  res$grp
})
write.csv(alpha_Continent_grp, row.names = FALSE,
          file = paste0(temp_out,"alpha_by_Continent_kw_aov_grp.csv"))
alpha_Continent_pvalue <- map_dfr(metric, function(a){
  res <- comp_kw_aov(wdat = dat, x = "Continent", y = a)
  res$comp$Metrics = a
  res$comp
})
alpha_Continent_pvalue$KW_pvalue_adj <- p.adjust(alpha_Continent_pvalue$KW_pvalue)
alpha_Continent_pvalue$aov_pvalue_adj <- p.adjust(alpha_Continent_pvalue$aov_pvalue)
write.csv(alpha_Continent_pvalue, row.names = FALSE,
          file = paste0(temp_out,"alpha_by_Continent_kw_aov_pvalue.csv"))  

alpha_Continent_summary <- dat %>% dplyr::select(Faith_PD:Pielou_evenness, Continent) %>% 
  group_by(Continent) %>% 
  summarise_all(my_summarize) %>% 
  ungroup() %>%
  pivot_longer(-Continent,
               names_to = "Metrics",
               values_to = "value") %>% 
  unnest(cols = value) %>% 
  left_join(alpha_Continent_grp, by = c("Continent"="Var", "Metrics"="Metrics"))

wdat <- dat %>% 
  dplyr::select(SampleID, Continent, Faith_PD:Pielou_evenness) %>% 
  pivot_longer(Faith_PD:Pielou_evenness, names_to = "Metrics", values_to = "value") 

g <- ggplot(wdat, aes(Continent, value)) + 
  geom_boxplot(aes(color = Continent, fill = Continent),alpha = 0.7,  outlier.alpha = 0) + 
  geom_point(aes(color = Continent), position = position_jitterdodge()) + 
  geom_text(data = alpha_Continent_summary, 
            aes(x =  Continent, y = 1.05 * Max, label = KW_group)) + 
  scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
  #scale_x_discrete(labels=country_labels) + 
  labs(x = "", y = "") +
  facet_wrap(~Metrics, nrow = 2, scales = "free") +
  theme_bw() + theme(#aspect.ratio = 0.8,
    legend.position="none",
    panel.background = element_rect(fill = NA),
    strip.text = element_text(size=12,color="black",face='bold'),
    axis.title = element_text(size=12,color="black",face='bold'),
    axis.text = element_text(size=12,color="black",face='bold'),
    axis.text.x = element_markdown(size=12, face='bold',vjust = 0.5, hjust = 0.5),
    axis.text.y = element_text(size=12,color="black",face='bold'),
    legend.text = element_text(size=12,color="black",face='bold'),
    legend.title = element_text(size=12,color="black",face='bold'),
    title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out, "alpha_by_Continent_full.pdf"), device = cairo_pdf, 
       plot = g, width = 14, height = 10, units = "in", dpi = 300) 

stat.test <- wdat %>% group_by(Metrics) %>% 
  rstatix::wilcox_test(value ~ Continent) %>% 
  rstatix::adjust_pvalue(method = "fdr") %>%
  rstatix::add_significance("p.adj") %>% 
  rstatix::add_xy_position(x = "Continent", dodge = 0.8, scales = "free_y") %>% 
  filter(p.adj.signif != "ns")

g <- ggplot(wdat, aes(Continent, value)) + 
  geom_boxplot(aes(color = Continent, fill = Continent),alpha = 0.7,  outlier.alpha = 0) + 
  geom_point(aes(color = Continent), position = position_jitterdodge()) + 
  stat_pvalue_manual(stat.test, label = "p.adj.signif", tip.length = 0) +
  # geom_text(data = alpha_Continent_summary, 
  #           aes(x =  Continent, y = 1.05 * Max, label = KW_group)) + 
  scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
  #scale_x_discrete(labels=country_labels) + 
  labs(x = "", y = "") +
  facet_wrap(~Metrics, nrow = 2, scales = "free") +
  theme_bw() + theme(#aspect.ratio = 0.8,
    legend.position="none",
    panel.background = element_rect(fill = NA),
    strip.text = element_text(size=12,color="black",face='bold'),
    axis.title = element_text(size=12,color="black",face='bold'),
    axis.text = element_text(size=12,color="black",face='bold'),
    axis.text.x = element_markdown(size=12, face='bold',vjust = 0.5, hjust = 0.5),
    axis.text.y = element_text(size=12,color="black",face='bold'),
    legend.text = element_text(size=12,color="black",face='bold'),
    legend.title = element_text(size=12,color="black",face='bold'),
    title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out, "alpha_by_Continent_full_pvalue.pdf"), device = cairo_pdf, 
       plot = g, width = 14, height = 10, units = "in", dpi = 300) 


```

## beta
### function
```{r}
load("~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/Global.Rdata")
library(usedist)
calcOmega2_dm <- function(x) {
  # Computes the effect size omega^2 parital using the output from adonis test
  # Args:
  #   x: adonis output
  # Returns:
  #   A dataframe with calculated effect size omega^2 partial added to original dataframe
  require(dplyr)
  N_t = x["Total", ]$Df + 1
  MSe = x["Residual", ]$SumOfSqs/x["Residual", ]$Df
  out = x %>% as.data.frame %>% rownames_to_column(var = "Variable") %>% mutate(Omega2_partial = ifelse(is.na(`F`), NA, (SumOfSqs - Df * MSe)/(SumOfSqs + (N_t - Df) * MSe)))
  return(out)
}
```

### stat
```{r}
temp_out <- "~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/output/global/beta/"
for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd, "_global"))
  wmeta = meta_global %>% filter(SampleID %in% labels(wdist))
  wdist <- dist_subset(wdist, wmeta$SampleID) 
  print(dd)
  # ## Country
  # disper <- betadisper(wdist, wmeta$Country)
  # disper_ova <- anova(disper)
  # disper_HSD <- TukeyHSD(disper)
  # write.csv(disper_ova, file = paste0(temp_out,"disper_anova_", dd, "_Country.csv"), quote = F)
  # write.csv(disper_HSD$group, file = paste0(temp_out,"disper_HSD_", dd, "_Country.csv"), quote = F)
  # 
  # per <- adonis2(wdist ~ Country, data = wmeta, by = "margin")
  # res <-  calcOmega2_dm(per)
  # write.csv(res, file = paste0(temp_out,"adonis_test_", dd, "_Country.csv"), quote = F, row.names = F)
  # 
  # ## Continent
  # disper <- betadisper(wdist, wmeta$Continent)
  # disper_ova <- anova(disper)
  # disper_HSD <- TukeyHSD(disper)
  # write.csv(disper_ova, file = paste0(temp_out,"disper_anova_", dd, "_Continent.csv"), quote = F)
  # write.csv(disper_HSD$group, file = paste0(temp_out,"disper_HSD_", dd, "_Continent.csv"), quote = F)
  # 
  # per <- adonis2(wdist ~ Continent, data = wmeta, by = "margin")
  # res <-calcOmega2_dm(per)
  # write.csv(res, file = paste0(temp_out,"adonis_test_", dd, "_Continent.csv"), quote = F, row.names = F)
  
  # ## postpartum 
  # wwmeta <- wmeta %>% filter(!is.na(Time_post_partum_day))
  # wwdist <- dist_subset(wdist, wwmeta$SampleID )
  # disper <- betadisper(wwdist, wwmeta$Time_post_partum_day)
  # disper_ova <- anova(disper)
  # disper_HSD <- TukeyHSD(disper)
  # write.csv(disper_ova, file = paste0(temp_out,"disper_anova_", dd, "_postpartum.csv"), quote = F)
  # write.csv(disper_HSD$group, file = paste0(temp_out,"disper_HSD_", dd, "_postpartum.csv"), quote = F)
  # 
  # per <- adonis2(wwdist ~ Time_post_partum_day, data = wwmeta, by = "margin")
  # res <-calcOmega2_dm(per)
  # write.csv(res, file = paste0(temp_out,"adonis_test_", dd, "_postpartum.csv"), quote = F, row.names = F)
  
  #   ## Maternal_age
  # wwmeta <- wmeta %>% filter(!is.na(Maternal_age))
  # wwdist <- dist_subset(wdist, wwmeta$SampleID )
  # disper <- betadisper(wwdist, wwmeta$Maternal_age)
  # disper_ova <- anova(disper)
  # disper_HSD <- TukeyHSD(disper)
  # write.csv(disper_ova, file = paste0(temp_out,"disper_anova_", dd, "_Maternal_age.csv"), quote = F)
  # write.csv(disper_HSD$group, file = paste0(temp_out,"disper_HSD_", dd, "_Maternal_age.csv"), quote = F)
  # 
  # per <- adonis2(wwdist ~ Maternal_age, data = wwmeta, by = "margin")
  # res <-calcOmega2_dm(per)
  # write.csv(res, file = paste0(temp_out,"adonis_test_", dd, "_Maternal_age.csv"), quote = F, row.names = F)
  # 
  # ## Maternal_weight_kg
  # wwmeta <- wmeta %>% filter(!is.na(Maternal_weight_kg))
  # wwdist <- dist_subset(wdist, wwmeta$SampleID )
  # disper <- betadisper(wwdist, wwmeta$Maternal_weight_kg)
  # disper_ova <- anova(disper)
  # disper_HSD <- TukeyHSD(disper)
  # write.csv(disper_ova, file = paste0(temp_out,"disper_anova_", dd, "_Maternal_weight_kg.csv"), quote = F)
  # write.csv(disper_HSD$group, file = paste0(temp_out,"disper_HSD_", dd, "_Maternal_weight_kg.csv"), quote = F)
  # 
  # per <- adonis2(wwdist ~ Maternal_weight_kg, data = wwmeta, by = "margin")
  # res <-calcOmega2_dm(per)
  # write.csv(res, file = paste0(temp_out,"adonis_test_", dd, "_Maternal_weight_kg.csv"), quote = F, row.names = F) 
  #   ## Birth_mode
  # wwmeta <- wmeta %>% filter(!is.na(Birth_mode))
  # wwdist <- dist_subset(wdist, wwmeta$SampleID )
  # disper <- betadisper(wwdist, wwmeta$Birth_mode)
  # disper_ova <- anova(disper)
  # disper_HSD <- TukeyHSD(disper)
  # write.csv(disper_ova, file = paste0(temp_out,"disper_anova_", dd, "_Birth_mode.csv"), quote = F)
  # write.csv(disper_HSD$group, file = paste0(temp_out,"disper_HSD_", dd, "_Birth_mode.csv"), quote = F)
  # 
  # per <- adonis2(wwdist ~ Birth_mode, data = wwmeta, by = "margin")
  # res <-calcOmega2_dm(per)
  # write.csv(res, file = paste0(temp_out,"adonis_test_", dd, "_Birth_mode.csv"), quote = F, row.names = F)
  
  ## all
  wwmeta <- wmeta %>% filter(!is.na(Birth_mode), !is.na(Maternal_weight_kg),!is.na(Maternal_age), !is.na(Time_post_partum_day))
  wwdist <- dist_subset(wdist, wwmeta$SampleID )
  # disper <- betadisper(wwdist, wwmeta$Birth_mode)
  # disper_ova <- anova(disper)
  # disper_HSD <- TukeyHSD(disper)
  # write.csv(disper_ova, file = paste0(temp_out,"disper_anova_", dd, "_Birth_mode.csv"), quote = F)
  # write.csv(disper_HSD$group, file = paste0(temp_out,"disper_HSD_", dd, "_Birth_mode.csv"), quote = F)
  
  per <- adonis2(wwdist ~ Birth_mode + Maternal_weight_kg + Maternal_age +  Time_post_partum_day + Country, data = wwmeta, by = "margin")
  res <-calcOmega2_dm(per)
  write.csv(res, file = paste0(temp_out,"adonis_test_", dd, "_All_Country.csv"), quote = F, row.names = F)
  
  per <- adonis2(wwdist ~ Birth_mode + Maternal_weight_kg + Maternal_age +  Time_post_partum_day + Continent, data = wwmeta, by = "margin")
  res <-calcOmega2_dm(per)
  write.csv(res, file = paste0(temp_out,"adonis_test_", dd, "_All_Continent.csv"), quote = F, row.names = F)
  
  

}
```

### PCoA
```{r}
library(gridExtra)
library(ggrepel)
for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd, "_global"))
  wmeta = meta_global %>% filter(SampleID %in% labels(wdist))
  wdist <- dist_subset(wdist, wmeta$SampleID) 
  
  ## PCoA
  wpcoa <- pcoa(wdist)
  varPC1 <- round(wpcoa$values$Relative_eig[1]*100, 2)
  varPC2 <- round(wpcoa$values$Relative_eig[2]*100, 2)
  x_lab <- paste0("PC1 (", varPC1, "%)")
  y_lab <- paste0("PC2 (", varPC2, "%)")
  wdat_pcoa <- merge(wpcoa$vectors[,1:2], 
                     wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = Axis.1, PC2 = Axis.2)
  
  ## Country
  wdat_center <- wdat_pcoa %>% group_by(Country) %>% 
    summarise(across(starts_with("PC"), mean)) %>% dplyr::select(-PCRConc) %>% 
    left_join(wmeta %>% dplyr::select(Country, Continent) %>% unique(),by = "Country")
  wdat <- wdat_pcoa %>% left_join(wdat_center %>% transmute(PC1.c = PC1, PC2.c = PC2, Country), by = "Country")
  
  g_tab <- read.csv(paste0(temp_out,"adonis_test_", dd, "_Country.csv")) %>% 
    filter(!is.na(Omega2_partial)) %>% 
    transmute(Name = Variable, Df = Df, SumOfSqs = round(SumOfSqs, 3),
              R2 = round(R2, 3), `F` = round(`F`, 3), Pvalue = round(`Pr..F.`,3),
              Omega2_partial = round(Omega2_partial,3)) %>% column_to_rownames(var = "Name")
  ### points
  g <- ggplot(wdat, aes(x = PC1, y = PC2, color = Country)) + 
    geom_point(size = 2) + 
    scale_color_manual(values = Country_color) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
      panel.background = element_rect(fill = NA),
      strip.text = element_text(size=12,color="black",face='bold'),
      axis.title = element_text(size=12,color="black",face='bold'),
      axis.text = element_text(size=12,color="black",face='bold'),
      axis.text.x = element_text(size=12,color="black",face='bold'),
      axis.text.y = element_text(size=12,color="black",face='bold'),
      legend.text = element_text(size=12,color="black",face='bold'),
      legend.title = element_text(size=12,color="black",face='bold'),
      title = element_text(size=12,color="black",face='bold'))  
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin)   
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_by_Country.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)  
  ### center
  g <- ggplot(wdat_center, aes(x = PC1, y = PC2, color = Continent)) + 
    geom_point(data = wdat, aes(x = PC1, y = PC2, color = Continent), size = 1, alpha = 0.5) + 
    geom_segment(data = wdat, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4,shape = 21, color = "black", aes(fill = Continent)) + 
    geom_text_repel(aes(label = Country), color = "black", max.overlaps = 20) + 
    scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
      panel.background = element_rect(fill = NA),
      strip.text = element_text(size=12,color="black",face='bold'),
      axis.title = element_text(size=12,color="black",face='bold'),
      axis.text = element_text(size=12,color="black",face='bold'),
      axis.text.x = element_text(size=12,color="black",face='bold'),
      axis.text.y = element_text(size=12,color="black",face='bold'),
      legend.text = element_text(size=12,color="black",face='bold'),
      legend.title = element_text(size=12,color="black",face='bold'),
      title = element_text(size=12,color="black",face='bold')) 
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin)    
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_by_Country_center.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300) 
  
  ## Continent
  wdat_center2 <- wdat_pcoa %>% group_by(Continent) %>% 
    summarise(across(starts_with("PC"), mean)) %>% dplyr::select(-PCRConc) %>% 
    left_join(wmeta %>% dplyr::select(Continent) %>% unique(),by = "Continent")
  wdat2 <- wdat_pcoa %>% left_join(wdat_center2 %>% transmute(PC1.c = PC1, PC2.c = PC2, Continent), by = "Continent")
  
  g_tab <- read.csv(paste0(temp_out,"adonis_test_", dd, "_Continent.csv")) %>% 
    filter(!is.na(Omega2_partial)) %>% 
    transmute(Name = Variable, Df = Df, SumOfSqs = round(SumOfSqs, 3),
              R2 = round(R2, 3), `F` = round(`F`, 3), Pvalue = round(`Pr..F.`,3),
              Omega2_partial = round(Omega2_partial,3)) %>% column_to_rownames(var = "Name")
  
  g <- ggplot(wdat_center2, aes(x = PC1, y = PC2, color = Continent)) + 
    geom_point(data = wdat2, aes(x = PC1, y = PC2, color = Continent), size = 1, alpha = 0.5) + 
    geom_segment(data = wdat2, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.5) + 
    geom_point(size = 4, shape = 21, color = "black", aes(fill = Continent)) + 
    geom_text_repel(aes(label = Continent), color = "black") + 
    scale_color_manual(values = Continent_color, aesthetics = c("color","fill")) + 
    #stat_ellipse(data = wdat2, aes(x = PC1, y = PC2, color = Continent)) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
      panel.background = element_rect(fill = NA),
      strip.text = element_text(size=12,color="black",face='bold'),
      axis.title = element_text(size=12,color="black",face='bold'),
      axis.text = element_text(size=12,color="black",face='bold'),
      axis.text.x = element_text(size=12,color="black",face='bold'),
      axis.text.y = element_text(size=12,color="black",face='bold'),
      legend.text = element_text(size=12,color="black",face='bold'),
      legend.title = element_text(size=12,color="black",face='bold'),
      title = element_text(size=12,color="black",face='bold'))
  xmin <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[1]
  xmax <- ggplot_build(g)$layout$panel_scales_x[[1]]$range$range[2]
  ymin <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[1]
  ymax <- ggplot_build(g)$layout$panel_scales_y[[1]]$range$range[2]
  g = g + ylim(c(ymin - (ymax - ymin)/5, ymax)) +
    annotation_custom(grob = tableGrob(g_tab, theme = ttheme_minimal(base_size = 10)),
                      xmin = xmin, xmax = xmax,
                      ymin = ymin - (ymax - ymin)/5, ymax = ymin)   
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_by_Continent_center.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)  
}
```

## taxa
### load data
```{r message=FALSE, warning=FALSE}
# load("~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/Global.Rdata")
# temp_out <- "~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/output/global/taxa/"
load("~/Dropbox/rutgers/20220606_RUN41_milk/Data/Global.Rdata")
temp_out <- "~/Dropbox/rutgers/20220606_RUN41_milk/Data/output/global/taxa/"
options(getClass.msg=FALSE)
phylo_global_s <- tax_glom(phylo_global_decontam_clean_rf, taxrank = "Species", NArm=FALSE)
phylo_global_g <- tax_glom(phylo_global_decontam_clean_rf, taxrank = "Genus", NArm=FALSE)
phylo_global_p <- tax_glom(phylo_global_decontam_clean_rf, taxrank = "Phylum", NArm=FALSE)

```

### unique taxa
```{r}
library(UpSetR)
for (taxrank in c("ASV","Species","Genus")){
  if (taxrank == "ASV"){
    ft <- otu_table(phylo_global_decontam_clean_rf@otu_table) %>% as.data.frame() %>% t()
  }else if(taxrank == "Species"){
    ft <- otu_table(phylo_global_s@otu_table) %>% as.data.frame() %>% t()
  }else if(taxrank == "Genus"){
    ft <- otu_table(phylo_global_g@otu_table) %>% as.data.frame() %>% t()
  }else{
    ft <- otu_table(phylo_global_f@otu_table) %>% as.data.frame() %>% t()
  }
  ft <- ft[meta_global_rf$SampleID,]
  asv_sum_country <- aggregate(ft, by = list(meta_global_rf$Country), sum)
  asv_sum_country_lst <- split(asv_sum_country[,-1], asv_sum_country[,1])
  asv_lst <- lapply(asv_sum_country_lst, function(dat){
    names(dat)[(which(dat[1,] > 1))]
  })
  g <- upset(fromList(asv_lst), 
             sets = c("Equatorial Guinea", "Kenya", "Senegal", "South Sudan", "Tanzania",
                      "Ecuador", "ElSalvador", "Mexico", "Peru", "Puerto Rico", "United States",
                      "Austria", "Germany", "Netherlands", "Norway", "Spain"),
             order.by = "freq", keep.order = TRUE,
             nsets = length(asv_lst),
             mainbar.y.label = paste0(taxrank, " Intersections"),
             sets.x.label = paste0(taxrank, " number by country"),
             nintersects = 40)  
  pdf(file = paste0(temp_out,"rarefied_sharing_", taxrank, ".pdf"), 
      width = 9, height = 7, useDingbats = F)
  print(g)
  dev.off() 
}

```

### phylums by country pie
```{r}
library(RColorBrewer)
col_vector = c(brewer.pal(12, "Set3"), brewer.pal(12, "Paired"), 
               brewer.pal(8, "Set2"), brewer.pal(7, "Accent")) %>% unique()

phylo_global_p <- tax_glom(phylo_global_decontam_clean_rf, taxrank = "Phylum", NArm=FALSE)
ft <- otu_table(phylo_global_p@otu_table) %>% as.data.frame() %>% t()
ft <- ft[meta_global_rf$SampleID,]

ft_ra <- ft/rowSums(ft)
ft_ra_country <- aggregate(ft_ra, by = list(meta_global_rf$Country), mean) %>%
  rename(Country = Group.1)

res <-  purrr::map_dfr(
  ft_ra_country %>% dplyr::select(-Country) %>% split(ft_ra_country$Country),
  function(dat){
    dat2 <- t(dat) %>% as.data.frame()
    names(dat2) <- "RA"
    dat3 <- dat2 %>% arrange(desc(RA)) %>% filter(RA > 0.01) %>% 
      rownames_to_column(var = "id")
    rbind(dat3, data.frame(id = "Other", RA = 1-sum(dat3$RA)))
  }, .id = "Country")

res <- res %>% left_join(taxa %>% dplyr::select(id, Phylum), by = "id") %>% 
  mutate(Taxa = ifelse(is.na(Phylum),"Other",Phylum))

temp <- res %>% dplyr::select(Taxa, RA) %>% group_by(Taxa) %>% dplyr::summarise(RA = mean(RA)) %>% arrange(desc(RA)) %>% filter(Taxa != "Other")

phylum_color = c(col_vector[1:(nrow(temp))],"Gray50")
names(phylum_color) = c(temp$Taxa, "Other")

dat_title <- meta_global_rf %>% count(Country) %>% 
  mutate(title = str_c(Country, " N=",n))

g_lst <- lapply(res$Country %>% unique(), function(x){
  title_x <- dat_title %>% filter(Country == x) %>% pull(title)
  g <- ggplot(res %>% filter(Country == x),
              aes(x = "", y = RA, fill = Taxa)) + 
    geom_bar(stat = "identity", width = 1, color = "white") + 
    scale_fill_manual(values = phylum_color, name = "") + 
    labs(title = title_x) + 
    coord_polar("y", start=0) + 
    theme_void() + theme(plot.title = element_text(face = "bold", size = 12, hjust = 0.5), legend.position = "none")  
})
names(g_lst) <- res$Country %>% unique()

lgplot <- ggplot(data.frame(x = factor(names(phylum_color), 
                                       levels = c(names(phylum_color)[1:length(phylum_color)-1] %>% sort(), "Other"), 
                                       ordered = T), 
                            y = 1), 
                 aes(x,y)) + 
  geom_col(aes(fill = x)) + 
  scale_color_manual(values = phylum_color, aesthetics = "fill", name = "")

lg <- as_ggplot(get_legend(lgplot))

p1 <- g_lst$`Equatorial Guinea` + g_lst$Kenya + g_lst$Senegal + g_lst$`South Sudan` + g_lst$Tanzania +
  g_lst$Ecuador + g_lst$ElSalvador + g_lst$Mexico + g_lst$Peru + g_lst$`Puerto Rico` + g_lst$`United States` + 
  g_lst$Austria + g_lst$Germany + g_lst$Netherlands + g_lst$Norway + g_lst$Spain +
  plot_layout(ncol = 4, guides = "collect")

layout <- "
ABCD#
EFGHQ
IJKLQ
MNOP#
"

p2 <- wrap_plots(A = g_lst$`Equatorial Guinea`, B = g_lst$Kenya, C = g_lst$Senegal, D = g_lst$`South Sudan`,
                 E = g_lst$Tanzania, F = g_lst$Ecuador, G = g_lst$ElSalvador, H = g_lst$Mexico,
                 I = g_lst$Peru, J = g_lst$`Puerto Rico`, K = g_lst$`United States`,  L = g_lst$Austria,
                 M = g_lst$Germany, N = g_lst$Netherlands, O = g_lst$Norway, P = g_lst$Spain, Q = lg,
                 design = layout)

ggsave(filename = paste0(temp_out,"top_phylum_pie.pdf"), device = cairo_pdf,
       plot = p2, width = 14, height = 10, units = "in", dpi = 300)
```

### heat map high pre
```{r}
library(ComplexHeatmap)

mat_col_fun = circlize::colorRamp2(c(-4, -3.3, -2.6, -1.9, -1.2, -0.5), 
                                   c("#253494","#2c7fb8","#41b6c4", "#7fcdbb", "#c7e9b4", "#ffffcc"))
col_fun = circlize::colorRamp2(c(-2, 0, 2), c("green", "white", "red"))

prev = 30
ra = 0.01

for (taxrank in c("ASV", "Species","Genus")){
  if(taxrank == "ASV"){
    ft <- otu_table(phylo_global_decontam_clean_rf@otu_table) %>% as.data.frame()
    taxa_temp <- taxa %>% mutate(taxa_long = str_c(Phylum,Species_l,ASV, sep = ";"))
  }else if(taxrank == "Species"){
    ft <- otu_table(phylo_global_s@otu_table) %>% as.data.frame()
    taxa_temp <- taxa %>% mutate(taxa_long = str_c(Phylum,Species_l, sep = ";"))
  }else if(taxrank == "Genus"){
    ft <- otu_table(phylo_global_g@otu_table) %>% as.data.frame()
    taxa_temp <- taxa %>% mutate(taxa_long = str_c(Phylum,Genus_l, sep = ";"))
  }
  n = nrow(meta_global_rf)
  ft2 <- t(ft)
  ft <- ft[rowSums(ft)>0, ]
  ft <- ft/colSums(ft)[1]
  dat <- data.frame(id = rownames(ft), prevalence = rowSums(ft>0)/n*100, abundance =  rowMeans(ft))
  dat <- dat %>% left_join(taxa, by = "id")
  ft2 <- ft2[meta_global_rf$SampleID,]
  asv_sum_country <- aggregate(ft2, by = list(meta_global_rf$Country), sum)
  asv_sum_country_lst <- split(asv_sum_country[,-1], asv_sum_country[,1])
  asv_lst <- lapply(asv_sum_country_lst, function(dat){
    names(dat)[(which(dat[1,] > 1))]
  })
  universal_id <- Reduce(intersect, asv_lst)
  dat <- dat %>% mutate(Universal = ifelse(id %in% universal_id, "Y", "N"))
  write.csv(dat, file = paste0(temp_out,"Prevalence_", taxrank, "_in_global.csv"), row.names = F, quote = F)
  ft2_lst <- split(as.data.frame(ft2), meta_global_rf$Country)
  res <- lapply(seq_along(ft2_lst), function(n){
    a <- ft2_lst[[n]]
    Country <- names(ft2_lst)[n]
    temp_pre <- colSums(a) /ncol(a) * 100 
    temp_dat <- data.frame(id = names(temp_pre), prevalence = temp_pre, abundance = colMeans(a)/(rowSums(a)[1]))
    temp_dat %>% top_n(n = 10, wt = prevalence) %>% mutate(Country = Country)
  })
  res_all <- do.call(rbind, res)
  write.csv(res_all, file = paste0(temp_out,"Prevalence_", taxrank, "_in_global_by_Country.csv"), row.names = F, quote = F)
}

for (taxrank in c("ASV", "Species", "Genus")){
  if(taxrank == "ASV"){
    ft <- otu_table(phylo_global_decontam_clean_rf@otu_table) %>% as.data.frame()
    taxa_temp <- taxa %>% mutate(taxa_long = str_c(Phylum,Species_l,ASV, sep = ";"))
  }else if(taxrank == "Species"){
    ft <- otu_table(phylo_global_s@otu_table) %>% as.data.frame()
    taxa_temp <- taxa %>% mutate(taxa_long = str_c(Phylum,Species_l, sep = ";"))
  }else if(taxrank == "Genus"){
    ft <- otu_table(phylo_global_g@otu_table) %>% as.data.frame()
    taxa_temp <- taxa %>% mutate(taxa_long = str_c(Phylum,Genus_l, sep = ";"))
  }
  
  select_top_pre <- read.csv(paste0(temp_out,"Prevalence_", taxrank, "_in_global_by_Country.csv"))
  top_pre_id <- select_top_pre$id %>% unique()
  dat <- read_csv(paste0(temp_out,"Prevalence_", taxrank, "_in_global.csv"))
  select_id <- dat %>% filter(prevalence>prev | abundance > ra) %>% pull(id)
  
  ft_logra <- log10((ft + 0.5)/colSums(ft)[1])
  
  meta_global_ord <- meta_global_rf %>% arrange(Continent, Country, SampleID)
  ft_select <- ft[select_id, meta_global_ord %>% pull(SampleID)]
  if (nrow(ft_select)>50){
    ft_select <- ft_select[(rowSums(ft_select) %>% sort(.,decreasing = T))[1:50] %>% names(),]
  }
  
  wft <- ft_logra[rownames(ft_select),]
  wft <- wft %>% rownames_to_column(var="id") %>% 
    left_join(taxa_temp %>% dplyr::select(id, taxa_long), by = "id") %>% 
    dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
  wft2 = wft %>% column_to_rownames(var = "taxa_long")
  wft2 <- wft2[,meta_global_ord$SampleID]
  df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
  df_col = meta_global_ord %>% dplyr::select(Continent, Country)
  column_ha = HeatmapAnnotation(df = df_col %>% dplyr::select(Continent, Country), 
                                col = list(Continent = Continent_color,
                                           Country = Country_color),
                                simple_anno_size = unit(8, "mm"), 
                                annotation_label = c("Continent", "Country"),
                                annotation_name_gp = gpar(fontsize = 14,fontface = "bold"), 
                                annotation_name_side = "left",
                                show_legend = F)
  row_dend = as.dendrogram(hclust(dist(wft2)))
  p1 = Heatmap(as.matrix(wft2), 
               col = mat_col_fun, 
               #name = "Relative Abundance",
               show_heatmap_legend = F,
               cluster_columns = F, show_column_names = F,
               #column_split = df_col$SamplingAge_group,
               cluster_rows = row_dend,
               row_names_gp = gpar(fontsize = 14),
               row_labels = df_row_name_label,
               row_dend_width = unit(15, "mm"),
               row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
               top_annotation = column_ha,
               height = unit(8, "mm")*dim(wft)[1],
               width = unit(0.4, "mm")*dim(wft)[2]
  )
  lgd1 = Legend(labels = names(Continent_color), title = "Continent", 
                legend_gp = gpar(fill = Continent_color),
                title_position = "topleft", 
                labels_gp = gpar(fontsize = 14, fontface = "bold"),
                title_gp = gpar(fontsize = 14, fontface = "bold"))
  lgd2 = Legend(labels = names(Country_color), title = "Country", 
                legend_gp = gpar(fill = Country_color),
                title_position = "topleft", 
                labels_gp = gpar(fontsize = 14, fontface = "bold"),
                title_gp = gpar(fontsize = 14, fontface = "bold"))
  lgd3 = Legend(col_fun = mat_col_fun, title = "Relative abundance", 
                at = c(-4, -3.3, -2.6, -1.9, -1.2, -0.5),                 
                labels = c("0.01%", "0.05%", "0.25%", "1.25%", "6.25%", "31.25%"), 
                title_position = "topleft",
                labels_gp = gpar(fontsize = 14, fontface = "bold"), 
                title_gp = gpar(fontsize = 14, fontface = "bold"))

  pd = packLegend(list = list(lgd1, lgd2, lgd3), direction = "vertical")  
  p = draw(p1, annotation_legend_side = "left", annotation_legend_list = pd)
  pdf(file = paste0(temp_out,"heatmap_pre_", prev, "_abund_", ra, "_country_", taxrank, "_supervised.pdf"), 
      width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
      height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
  draw(p)
  dev.off()  
  
  p2 = Heatmap(as.matrix(wft2), 
               col = mat_col_fun, 
               show_heatmap_legend = F,
               cluster_columns = T, show_column_names = F,
               cluster_rows = row_dend,
               row_names_gp = gpar(fontsize = 14),
               row_labels = df_row_name_label,
               row_dend_width = unit(15, "mm"),
               row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
               top_annotation = column_ha,
               height = unit(8, "mm")*dim(wft)[1],
               width = unit(0.4, "mm")*dim(wft)[2]
  )  
  p = draw(p2, annotation_legend_side = "left", annotation_legend_list = pd)
  pdf(file = paste0(temp_out,"heatmap_pre_", prev, "_abund_", ra, "_country_", taxrank, "_unsupervised.pdf"), 
      width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
      height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
  draw(p)
  dev.off()
  
  ft_select <- ft[top_pre_id, meta_global_ord %>% pull(SampleID)]
  if (nrow(ft_select)>50){
    ft_select <- ft_select[(rowSums(ft_select) %>% sort(.,decreasing = T))[1:50] %>% names(),]
  }
  wft <- ft_logra[rownames(ft_select),]
  wft <- wft %>% rownames_to_column(var="id") %>% 
    left_join(taxa_temp %>% dplyr::select(id, taxa_long), by = "id") %>% 
    dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
  wft2 = wft %>% column_to_rownames(var = "taxa_long")
  wft2 <- wft2[,meta_global_ord$SampleID]
  df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
  row_dend = as.dendrogram(hclust(dist(wft2)))
  
  p3 = Heatmap(as.matrix(wft2), 
               col = mat_col_fun, 
               #name = "Relative Abundance",
               show_heatmap_legend = F,
               cluster_columns = F, show_column_names = F,
               #column_split = df_col$SamplingAge_group,
               cluster_rows = row_dend,
               row_names_gp = gpar(fontsize = 14),
               row_labels = df_row_name_label,
               row_dend_width = unit(15, "mm"),
               row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
               top_annotation = column_ha,
               height = unit(8, "mm")*dim(wft)[1],
               width = unit(0.4, "mm")*dim(wft)[2]
  )  
  p = draw(p3, annotation_legend_side = "left", annotation_legend_list = pd)
  pdf(file = paste0(temp_out,"heatmap_top_pre_by_country_", taxrank, "_supervised.pdf"), 
      width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
      height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
  draw(p)
  dev.off()    
  
  p4 = Heatmap(as.matrix(wft2), 
               col = mat_col_fun, 
               show_heatmap_legend = F,
               cluster_columns = T, show_column_names = F,
               cluster_rows = row_dend,
               row_names_gp = gpar(fontsize = 14),
               row_labels = df_row_name_label,
               row_dend_width = unit(15, "mm"),
               row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
               top_annotation = column_ha,
               height = unit(8, "mm")*dim(wft)[1],
               width = unit(0.4, "mm")*dim(wft)[2]
  )  
  p = draw(p4, annotation_legend_side = "left", annotation_legend_list = pd)
  pdf(file = paste0(temp_out,"heatmap_top_pre_by_country_", taxrank, "_unsupervised.pdf"), 
      width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
      height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
  draw(p)
  dev.off()  
}

```

### ANCOM by country
```{r}
library(ANCOMBC)
library(ComplexHeatmap)

for (taxrank in c("Genus", "ASV", "Species")){
  if(taxrank == "ASV"){
    phylotmp <- phylo_global_decontam_clean_rf
    taxa_temp <- taxa %>% mutate(taxa_long = str_c(Phylum,Species_l,ASV, sep = ";"))
  }else if(taxrank == "Species"){
    phylotmp <- phylo_global_s
    taxa_temp <- taxa %>% mutate(taxa_long = str_c(Phylum,Species_l, sep = ";"))
  }else if(taxrank == "Genus"){
    phylotmp <- phylo_global_g
    taxa_temp <- taxa %>% mutate(taxa_long = str_c(Phylum,Genus_l, sep = ";"))
  }
  #trace(ancombc2, edit = TRUE)
  #line 66 data = as.data.frame(meta_data)
  
  output <- ancombc2(data = phylotmp, 
                     #meta_data = meta_global_rf, 
                     tax_level = NULL,
                     fix_formula = "Country", 
                     rand_formula = NULL,
                     p_adj_method = "holm", pseudo_sens = TRUE,
                     prv_cut = 0.1, lib_cut = 1000, s0_perc = 0.05,
                     group = "Country", struc_zero = TRUE, neg_lb = FALSE,
                     alpha = 0.05, n_cl = 16, verbose = TRUE,
                     global = TRUE, pairwise = TRUE, 
                     dunnet = TRUE, trend = FALSE,
                     iter_control = list(tol = 1e-5, max_iter = 20, 
                                         verbose = FALSE),
                     em_control = list(tol = 1e-5, max_iter = 100),
                     lme_control = NULL, 
                     mdfdr_control = list(fwer_ctrl_method = "holm", B = 100), 
                     trend_control = NULL)
  
  res <- output$res_global %>% as.data.frame()
  res <- res %>% filter(diff_abn)
  if(nrow(res) <2 ){next}
  
  for (dtype in c("bias_correct_log", "log")){
    if (dtype == "bias_correct_log"){
      bias_correct_log_table <- output$bias_correct_log_table
      bias_correct_log_table[is.na(bias_correct_log_table)] <- 0
      ft <- bias_correct_log_table[res$taxon,]
      mat_col_fun = circlize::colorRamp2(c(-3, -2, -1, -0.5, 0.5 ,1, 2, 3), 
                                         c("#253494","#225ea8","#1d91c0","#41b6c4","#7fcdbb","#c7e9b4","#edf8b1","#ffffd9"))
      lgd3 = Legend(col_fun = mat_col_fun, title = "Relative abundance", 
                    at = c(-3, -1.5, -0.5, 0 ,0.5, 1.5, 3),                 
                    #labels = c("0.01%", "0.05%", "0.25%", "1.25%", "6.25%", "31.25%"), 
                    title_position = "topleft",
                    labels_gp = gpar(fontsize = 14, fontface = "bold"), 
                    title_gp = gpar(fontsize = 14, fontface = "bold"))
    }else if(dtype == "log"){
      mat_col_fun = circlize::colorRamp2(c(-4, -3.3, -2.6, -1.9, -1.2, -0.5), 
                                   c("#253494","#2c7fb8","#41b6c4", "#7fcdbb", "#c7e9b4", "#ffffcc"))
      ft <- otu_table(phylotmp@otu_table) %>% as.data.frame()
      ft <- log10((ft + 0.5)/colSums(ft)[1])
      ft <- ft[res$taxon,]
      lgd3 = Legend(col_fun = mat_col_fun, title = "Relative abundance", 
                    at = c(-4, -3.3, -2.6, -1.9, -1.2, -0.5),                 
                    labels = c("0.01%", "0.05%", "0.25%", "1.25%", "6.25%", "31.25%"), 
                    title_position = "topleft",
                    labels_gp = gpar(fontsize = 14, fontface = "bold"), 
                    title_gp = gpar(fontsize = 14, fontface = "bold"))
    }
    wft <- ft %>% rownames_to_column(var="id") %>% 
      left_join(taxa_temp %>% dplyr::select(id, taxa_long), by = "id") %>% 
      dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
    wft2 = wft %>% column_to_rownames(var = "taxa_long")
    df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
    meta_global_ord <- meta_global_rf %>% filter(SampleID %in% colnames(wft2)) %>%  arrange(Continent, Country, SampleID)
    wft2 <- wft2[,meta_global_ord$SampleID]
    df_col = meta_global_ord %>% dplyr::select(Continent, Country)
    column_ha = HeatmapAnnotation(df = df_col %>% dplyr::select(Continent, Country), 
                                  col = list(Continent = Continent_color,
                                             Country = Country_color),
                                  simple_anno_size = unit(8, "mm"), 
                                  annotation_label = c("Continent", "Country"),
                                  annotation_name_gp = gpar(fontsize = 14,fontface = "bold"), 
                                  annotation_name_side = "left",
                                  show_legend = F)
    row_dend = as.dendrogram(hclust(dist(wft2)))
    
    p1 = Heatmap(as.matrix(wft2), 
                 col = mat_col_fun, 
                 #name = "Relative Abundance",
                 show_heatmap_legend = F,
                 cluster_columns = F, show_column_names = F,
                 #column_split = df_col$SamplingAge_group,
                 cluster_rows = row_dend,
                 row_names_gp = gpar(fontsize = 14),
                 row_labels = df_row_name_label,
                 row_dend_width = unit(15, "mm"),
                 row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
                 top_annotation = column_ha,
                 height = unit(8, "mm")*dim(wft)[1],
                 width = unit(0.4, "mm")*dim(wft)[2]
    )
    lgd1 = Legend(labels = names(Continent_color), title = "Continent", 
                  legend_gp = gpar(fill = Continent_color),
                  title_position = "topleft", 
                  labels_gp = gpar(fontsize = 14, fontface = "bold"),
                  title_gp = gpar(fontsize = 14, fontface = "bold"))
    lgd2 = Legend(labels = names(Country_color), title = "Country", 
                  legend_gp = gpar(fill = Country_color),
                  title_position = "topleft", 
                  labels_gp = gpar(fontsize = 14, fontface = "bold"),
                  title_gp = gpar(fontsize = 14, fontface = "bold"))

    pd = packLegend(list = list(lgd1, lgd2, lgd3), direction = "vertical")  
    p = draw(p1, annotation_legend_side = "left", annotation_legend_list = pd)
    pdf(file = paste0(temp_out,"heatmap_diff_country_", taxrank,"_", dtype ,"_supervised.pdf"), 
        width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
        height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
    draw(p)
    dev.off()  
    
    p2 = Heatmap(as.matrix(wft2), 
                 col = mat_col_fun, 
                 show_heatmap_legend = F,
                 cluster_columns = T, show_column_names = F,
                 cluster_rows = row_dend,
                 row_names_gp = gpar(fontsize = 14),
                 row_labels = df_row_name_label,
                 row_dend_width = unit(15, "mm"),
                 row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
                 top_annotation = column_ha,
                 height = unit(8, "mm")*dim(wft)[1],
                 width = unit(0.4, "mm")*dim(wft)[2]
    )  
    p = draw(p2, annotation_legend_side = "left", annotation_legend_list = pd)
    pdf(file = paste0(temp_out,"heatmap_diff_country_", taxrank,"_", dtype , "_unsupervised.pdf"), 
        width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
        height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
    draw(p)
    dev.off()
  }
}

```

### decontam taxa
```{r}
load("~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/Global.Rdata")
temp_out <- "~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/output/global/taxa/"
contam_asv <- read_csv("~/Dropbox/Rutgers/20220606_RUN41_milk/Data/output/decontam_details.csv",show_col_types = FALSE)

ft <- otu_table(phylo_global@otu_table) %>% as.data.frame()
n = nrow(meta_global)
ft <- ft[rowSums(ft)>0, ]
ft <- ft/colSums(ft)[1]
dat <- data.frame(id = rownames(ft), prevalence = rowSums(ft>0)/n*100, abundance =  rowMeans(ft))

dat2 <- dat %>% filter(id %in% (contam_asv %>% filter(prev=="Y") %>% pull(id))) %>% 
  left_join(taxa, by = "id") %>% 
  dplyr::select(Kingdom:ASV, prevalence, abundance)

write_csv(dat2, file = paste0(temp_out,"contam_asv_pre.csv"))
```

### methanogen
```{r}
ft <- otu_table(phylo_global_g@otu_table) %>% as.data.frame() %>% 
  rownames_to_column(var = "id") %>% 
  left_join(taxa, by = "id") %>% 
  filter(str_detect(taxa_long, "Methanobacteriales")) %>% 
  dplyr::select(-c(id,Kingdom:ASV, taxa_long, Class_l:Family_l, Species_l)) %>% column_to_rownames(var = "Genus_l") %>% 
  t() %>% as.data.frame() %>% rownames_to_column(var = "SampleID") %>% 
  left_join(meta_global_rf %>% dplyr::select(SampleID, Country), by = "SampleID") 

wdat <- ft %>% 
  mutate(ra_g_Methanosphaera = g_Methanosphaera/depth,
         ra_g_Methanobacterium = g_Methanobacterium/depth,
         ra_g_Methanobrevibacter = g_Methanobrevibacter/depth,
         none = g_Methanosphaera + g_Methanobacterium + g_Methanobrevibacter) %>% 
  filter(none > 0)

write_csv(wdat, file = paste0(temp_out,"Methanogen_RA.csv"))

my_summarize2 <-  function(x){
  data.frame(
    N = n(),
    N.none0 = sum(x>0),
    # Min = min(x,na.rm = T),
    # Med = median(x, na.rm = T),
    Mean = mean(x, na.rm = T)
    # Max = max(x,na.rm = T),
    # SD = sd(x,na.rm = T)
  ) %>% 
    mutate(
      percent.none0 = N.none0/N
    )
}

methanogen_summary <- ft %>% dplyr::select(-SampleID) %>% 
      group_by(Country) %>% 
      summarise_all(my_summarize2) %>% 
      ungroup() %>% 
      unnest(cols = c("g_Methanosphaera","g_Methanobacterium","g_Methanobrevibacter"),names_sep = ";")

write_csv(methanogen_summary, file = paste0(temp_out,"Methanogen_summary.csv"))
```

### taxa diff boxplot by country
```{r}
diff_taxa <- c("g_Acinetobacter", "g_Alistipes", "g_Bacillus", "g_Bacteroides", 
               "g_Bifidobacterium", "g_Gemella", "g_Lactobacillus","g_Muribaculaceae",
               "g_Prevotella", "g_Pseudomonas", "g_Staphylococcus","g_Streptococcus")

gg = "g_Pseudomonas"

ft <- otu_table(phylo_global_g@otu_table) %>% as.data.frame()
ft_ra <- ft/colSums(ft)[1]
ft_ra <- ft_ra %>% rownames_to_column(var = "id") %>% left_join(taxa, by = "id")

ft_log <- log10((ft + 0.5)/colSums(ft)[1])
ft_log <- ft_log %>% rownames_to_column(var = "id") %>% left_join(taxa, by = "id")
##log is better
g_lst <- list()
for (gg in diff_taxa){
  for (dtype in c("log", "ra")){
    if(dtype == "log"){
      ft <- ft_log
    }else{
      ft <- ft_ra
    }
    ft <- ft %>% filter(Genus_l == gg) %>% 
      column_to_rownames(var = "id") %>% t() %>% as.data.frame() %>% 
      rownames_to_column(var = "SampleID") 
    names(ft) <- c("SampleID","Value")
    dat <- ft %>% left_join(meta_global_rf, by = "SampleID") %>% 
      mutate(Country = factor(Country,
                              levels = c("Equatorial Guinea", "Kenya", "Senegal", "South Sudan", "Tanzania",
                                         "Ecuador", "ElSalvador", "Mexico", "Peru", "Puerto Rico", "United States",
                                         "Austria", "Germany", "Netherlands", "Norway", "Spain"),
                              ordered = T),
             Value = as.numeric(Value)) %>% 
      filter(!is.na(Country))
    res <- comp_kw_aov(wdat = dat, x = "Country", y = "Value")
    res_summary <- dat %>% dplyr::select(Value, Country) %>% 
      group_by(Country) %>% 
      summarise_all(my_summarize) %>% 
      ungroup() %>%
      unnest(cols = Value) %>% 
      left_join(res$grp, by = c("Country"="Var")) %>% 
      mutate(N_lab = str_c("",N))
    
    g <- ggplot(dat %>% 
                  mutate(Country = factor(Country, levels= res_summary %>% arrange(desc(Med)) %>% pull(Country), ordered = T)),
                aes(Country, Value)) + 
      geom_boxplot(aes(color = Country, fill = Country), alpha = 0.6,  outlier.alpha = 0) + 
      geom_point(aes(color = Country), position = position_jitterdodge()) + 
      geom_text(data = res_summary, 
                aes(x =  Country, y = 1.15 * Max, label = KW_group)) +
      # geom_text(data = res_summary, 
      #           aes(x =  Country, y = -0.01, label = N_lab)) +
      scale_color_manual(values = Country_color, aesthetics = c("color","fill")) + 
      scale_x_discrete(labels=Country_labels) + 
      labs(x = "", y = ifelse(dtype == "log","LogRA","RA"), title = gg) +
      theme_bw() + theme(#aspect.ratio = 0.8,
        legend.position="none",
        panel.background = element_rect(fill = NA),
        strip.text = element_text(size=12,color="black",face='bold'),
        axis.title = element_text(size=12,color="black",face='bold'),
        axis.text = element_text(size=12,color="black",face='bold'),
        axis.text.x = element_markdown(size=12, face='bold',angle=90,vjust = 0.5, hjust = 1),
        axis.text.y = element_text(size=12,color="black",face='bold'),
        legend.text = element_text(size=12,color="black",face='bold'),
        legend.title = element_text(size=12,color="black",face='bold'),
        title = element_text(size=12,color="black",face='bold'))
    
    if(dtype == "log"){g_lst[[gg]] <- g}
    ggsave(filename = paste0(temp_out, gg,"_by_Country_", dtype, "_ordered.pdf"), device = cairo_pdf, 
           plot = g, width = 5, height = 5, units = "in", dpi = 300) 
  }
}

p <- g_lst[[1]]
for(i in c(2:length(g_lst))){
  p <- p + g_lst[[i]]
}
p + plot_layout(ncol = 4)
ggsave(filename = paste0(temp_out,"diff_genus_by_Country_log_ordered.pdf"), device = cairo_pdf, 
       plot = p, width = 20, height = 15, units = "in", dpi = 300) 




```
