==> code/08_new_global_network.Rmd <==
---
title: "Alpha diversity"
author: "Haipeng Sun"
date: "8/1/2022"
output: html_document
---

## library
```{r setup, message=FALSE, warning=FALSE}
library(phyloseq)
library(purrr)
library(furrr)
library(picante)
library(vegan)
library(MASS)
library(ggpubr)
library(effectsize)
library(ggtext)
library(glue)
library(SpiecEasi)
library(igraph)
library(tidygraph)
library(ggraph)
library(rstatix)
library(tidyverse)
```

## load data
```{r}
load("~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/Global_ITS.Rdata")
load("~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/Global.Rdata")
temp_out = "~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/output/global/network2/"
meta_both <- meta_both %>% filter(Project != "Covid Mom", SampleID != "BR34.1") 
```

## function
```{r}
aggregate_ft <- function(ft, taxa, L = "Genus"){
  #input: 
  #ft : feature in column matrix
  #taxa : df with id, taxa levels
  #L : taxa levels to aggregate c("Phylum","Class","Order", "Family", "Genus","Species")
  #output: feature in column matrix to a taxa level
  ft_agg <- ft %>% t() %>% as.data.frame() %>% 
    rownames_to_column(var = "id") %>% 
    left_join(taxa %>% select(id, Kingdom:Species), by = "id")
  if(L == "Species"){
    ft_agg <- ft_agg %>% mutate(grp = str_c(Kingdom, Phylum, Class, Order, Family, Genus, Species, sep = ";"))
  }else if(L == "Genus"){
    ft_agg <- ft_agg %>% mutate(grp = str_c(Kingdom, Phylum, Class, Order, Family, Genus, sep = ";"))
  }else if(L == "Family"){
    ft_agg <- ft_agg %>% mutate(grp = str_c(Kingdom, Phylum, Class, Order, Family, sep = ";"))
  }else if(L == "Order"){
    ft_agg <- ft_agg %>% mutate(grp = str_c(Kingdom, Phylum, Class, Order, sep = ";"))
  }else if(L == "Class"){
    ft_agg <- ft_agg %>% mutate(grp = str_c(Kingdom, Phylum, Class, sep = ";"))
  }else if(L == "Phylum"){
    ft_agg <- ft_agg %>% mutate(grp = str_c(Kingdom,Phylum, sep = ";"))
  }
  ft_agg <- ft_agg %>% select(-c(id,Kingdom:Species)) %>% 
    select(grp ,dplyr::everything()) %>% 
    group_by(grp) %>%
    summarise(across(everything(), sum),
              .groups = 'drop') %>% 
    ungroup() %>% 
    column_to_rownames(var = "grp") %>% t()
  ft_agg
}

filter_ra_pre <- function(ft_mt, ra_mt, abundance_cutoff = 0.001, prevalence_cutoff = 0.1){
  #input: feature count in column matrix
  #input: feature ra in column matrix
  #output: feature in column matrix filtered abundance and prevalence
  ft_filter <- ft_mt[, colMeans(ra_mt) > abundance_cutoff]
  ft_filter <- ft_filter[, colSums(ft_filter>0) > nrow(ft_filter)*prevalence_cutoff]  
  ft_filter
}

run_sparcc <- function(mt_wk, iter = 20, inner_iter = 10, th = 0.1, R = 100, ncpus = 10, seed = 1){
  
  scc0 <- sparcc(mt_wk, iter = iter, inner_iter = inner_iter, th = th)
  sparcc_cors = scc0$Cor
  colnames(sparcc_cors) = colnames(mt_wk)
  rownames(sparcc_cors) = colnames(mt_wk)
  
  set.seed(seed = seed)
  sparcc_boot_res <- sparccboot(mt_wk, R = R, ncpus = ncpus)
  sparcc_boot_pval <- pval.sparccboot(sparcc_boot_res, sided = "both")
  
  sparcc_pval <- diag(0.5, nrow = dim(sparcc_cors)[1], ncol = dim(sparcc_cors)[1])
  sparcc_pval[upper.tri(sparcc_pval, diag=FALSE)] <- sparcc_boot_pval$pvals
  sparcc_pval <- sparcc_pval + t(sparcc_pval)
  colnames(sparcc_pval) = colnames(mt_wk)
  rownames(sparcc_pval) = colnames(mt_wk)
  
  sparcc_boot_cors <- diag(0.5, nrow = dim(sparcc_cors)[1], ncol = dim(sparcc_cors)[1])
  sparcc_boot_cors[upper.tri(sparcc_boot_cors, diag=FALSE)] <- sparcc_boot_pval$cors
  sparcc_boot_cors <- sparcc_boot_cors + t(sparcc_boot_cors)
  colnames(sparcc_boot_cors) = colnames(mt_wk)
  rownames(sparcc_boot_cors) = colnames(mt_wk)
  
  res = list(cor = sparcc_boot_cors, pval = sparcc_pval)
  res
}

tidy_graph_edge_node <- function(cor_tidy){
  ## help function to get tidy graph
  ## filter any unclear taxa
  n_level <- str_count(cor_tidy$from[1], ";") + 1
  taxa_level <- c("Kingdom","Phylum","Class","Order", "Family", "Genus","Species")
  gg <- cor_tidy %>% 
    mutate(weight = abs(cor),
           from_name = str_split_i(from, ";",n_level),
           to_name = str_split_i(to, ";",n_level)) %>% 
    mutate(from_name = str_remove(from_name, "g_"),
           to_name = str_remove(to_name, "g_")) %>% 
    filter(from_name != "", to_name != "") %>% 
    filter(!str_detect(from_name,"_|-"), !str_detect(to_name,"_|-")) %>% 
    filter(!str_detect(from_name, "[0-9]"), !str_detect(to_name, "[0-9]")) %>% 
    filter(!str_detect(from_name, "uncultured"), !str_detect(to_name, "uncultured")) %>% 
    as_tbl_graph(directed = F)  
  
  edge_list <-gg %>%
    activate(edges) %>%
    data.frame() %>% 
    mutate(col = ifelse(cor > 0, "pos", "neg"))
  node_list <-gg %>%
    activate(nodes) %>%
    data.frame() 
  node_name <- str_split_fixed(node_list$name,";" ,n_level) %>% 
    as.data.frame() 
  names(node_name) <- taxa_level[1:n_level]
  node_name <- node_name %>% 
    mutate(label2 = str_c(str_remove(Kingdom, "k_"), Genus,sep = ";"),
           label3 = str_remove(Genus, "g_"),
           col_k = str_remove(Kingdom, "k_"))
  node_list <- bind_cols(node_list, node_name) 
  
  res <- list(node = node_list, edge = edge_list)
}

calcOmega2_dm <- function(x) {
    # Computes the effect size omega^2 parital using the output from adonis test
    # Args:
    #   x: adonis output
    # Returns:
    #   A dataframe with calculated effect size omega^2 partial added to original dataframe
    require(dplyr)
    N_t = x["Total", ]$Df + 1
    MSe = x["Residual", ]$SumOfSqs/x["Residual", ]$Df
    out = x %>% as.data.frame %>% rownames_to_column(var = "Variable") %>% mutate(Omega2_partial = ifelse(is.na(`F`), NA, (SumOfSqs - Df * MSe)/(SumOfSqs + (N_t - Df) * MSe)))
    return(out)
}

comp_kw_aov = function(wdat, x, y){
  #input data, respond variable y, categorize variable x
  require(agricolae)
  ktest1 = kruskal(wdat[y], wdat[x], alpha = 0.05, p.adj="none", group=T, main = "wdat", console=FALSE)
  out_g_1 = ktest1$groups %>% rownames_to_column(var = "Var") %>% dplyr::select(Var, groups) %>% rename(KW_group = groups)
  ktest2 = kruskal(wdat[y], wdat[x], alpha = 0.05, p.adj="none", group=F, main = "wdat", console=FALSE)
  out_c_1 = ktest2$comparison %>% rownames_to_column(var = "comp") %>% dplyr::select(comp, pvalue) %>% rename(KW_pvalue = pvalue)
  fit = aov(as.formula(paste0(y, "~", x)), wdat)
  tukey1 = HSD.test(fit, x, group = T)
  out_g_2 = tukey1$groups %>% rownames_to_column(var = "Var") %>% dplyr::select(Var, groups) %>% rename(aov_group = groups)
  tukey2 = HSD.test(fit, x, group = F)
  out_c_2 = tukey2$comparison %>% rownames_to_column(var = "comp") %>% dplyr::select(comp, pvalue) %>% rename(aov_pvalue = pvalue)
  out_g = merge(out_g_1, out_g_2,by = 1)
  out_c = merge(out_c_1, out_c_2,by = 1)
  res = list(out_g, out_c)
  names(res) = c("grp","comp")
  res
}

my_summarize = function(x){
  data.frame(
             N = n(),
             Min = min(x,na.rm = T),
             Q5 = quantile(x, probs = 0.05, na.rm = T),
             Q10 = quantile(x, probs = 0.1, na.rm = T),
             Q25 = quantile(x, probs = 0.25, na.rm = T),
             Med = median(x, na.rm = T),
             Mean = mean(x, na.rm = T),
             Q75 = quantile(x, probs = 0.75, na.rm = T),
             Q95 = quantile(x, probs = 0.95, na.rm = T),
             Max = max(x,na.rm = T)
             )
}
```

## network
### data
```{r}
# ITS
ft_its_wk <- ft_ITS %>% t() %>% 
  as.data.frame() %>% rownames_to_column(var = "SampleID_clean") %>% 
  left_join(meta_global %>% select(SampleID, SampleID_clean, Country), by = "SampleID_clean") %>% 
  filter(!is.na(Country)) %>% 
  select(-c(SampleID, Country)) %>% 
  column_to_rownames(var = "SampleID_clean")
  
ft_its_wk <- ft_its_wk[,colSums(ft_its_wk)>0]
ft_its_wk <- ft_its_wk[rowSums(ft_its_wk)>0,]
ra_its_wk <- sweep(ft_its_wk, 1, rowSums(ft_its_wk), `/`)

## genus
ft_its_wk_g <- aggregate_ft(ft = ft_its_wk, taxa = taxa_ITS, L = "Genus")
ft_its_wk_g <- ft_its_wk_g[,colSums(ft_its_wk_g)>0]
ra_its_wk_g <- sweep(ft_its_wk_g, 1, rowSums(ft_its_wk_g), `/`)


# 16S
ft_16s_wk <- phylo_global_decontam_clean_rf@otu_table@.Data %>% t() %>% 
  as.data.frame() %>% rownames_to_column(var = "SampleID") %>% 
  mutate(SampleID_clean = str_remove_all(SampleID, " ")) %>% 
  filter(SampleID_clean %in% (meta_global %>% pull(SampleID_clean))) %>% 
  select(-SampleID) %>% 
  column_to_rownames(var = "SampleID_clean")

ft_16s_wk <- ft_16s_wk[,colSums(ft_16s_wk)>0]
ft_16s_wk <- ft_16s_wk[rowSums(ft_16s_wk)>0,]
ra_16s_wk <- sweep(ft_16s_wk, 1, rowSums(ft_16s_wk), `/`)

## genus
ft_16s_wk_g <- aggregate_ft(ft = ft_16s_wk, taxa = taxa, L = "Genus")
ft_16s_wk_g <- ft_16s_wk_g[,colSums(ft_16s_wk_g)>0]
ra_16s_wk_g <- sweep(ft_16s_wk_g, 1, rowSums(ft_16s_wk_g), `/`)
```


### use sparcc ab 0.01, pre 0.1, cor 0.3,
```{r}
abundance_cutoff <- 0.01
prevalence_cutoff <- 0.1

ft_its_wk_g_filter <- filter_ra_pre(ft_mt = ft_its_wk_g, ra_mt = ra_its_wk_g, abundance_cutoff = abundance_cutoff, prevalence_cutoff = prevalence_cutoff)
ft_its_wk_g_filter <- ft_its_wk_g_filter  %>% as.data.frame() %>% 
  rownames_to_column(var = "SampleID_clean") %>% arrange(SampleID_clean) %>% filter(SampleID_clean != "BR34.1")

ft_16s_wk_g_filter <- filter_ra_pre(ft_mt = ft_16s_wk_g, ra_mt = ra_16s_wk_g, abundance_cutoff = abundance_cutoff, prevalence_cutoff = prevalence_cutoff)
ft_16s_wk_g_filter <- ft_16s_wk_g_filter %>% as.data.frame() %>% 
  rownames_to_column(var = "SampleID_clean") %>% arrange(SampleID_clean) %>% filter(SampleID_clean != "BR34.1")

res <- list()

mt_wk <- ft_16s_wk_g_filter %>% column_to_rownames(var = "SampleID_clean") %>% as.matrix()
set.seed(2025)
res[["16S All"]] <- run_sparcc(mt_wk = mt_wk, R = 1000)

for(ct in c("AFRICA","AMERICA","EUROPE")){
  mt_wkk <- mt_wk %>% as.data.frame() %>% rownames_to_column(var = "SampleID_clean") %>% 
    filter(SampleID_clean %in% (meta_global %>% filter(Continent == ct) %>% pull(SampleID_clean))) %>% 
    column_to_rownames(var = "SampleID_clean") %>% as.matrix()
  set.seed(2025)
  res[[paste0("16S ",ct)]] <- run_sparcc(mt_wk = mt_wkk, R = 1000)
}  

aa <- ft_16s_wk_g_filter %>% filter(SampleID_clean %in% ft_its_wk_g_filter$SampleID_clean) %>% 
  column_to_rownames(var = "SampleID_clean")
aa <- aa[,colSums(aa)>0] %>% 
  rownames_to_column(var = "SampleID_clean")
mt_wk <- ft_its_wk_g_filter %>% 
  left_join(aa, by = "SampleID_clean")  %>% 
  column_to_rownames(var = "SampleID_clean") %>% 
  as.matrix()
set.seed(2025)
res[["Both All"]] <- run_sparcc(mt_wk = mt_wk, R = 1000)

for(ct in c("AFRICA", "AMERICA")){
  mt_wkk <- mt_wk %>% as.data.frame() %>% rownames_to_column(var = "SampleID_clean") %>% 
    filter(SampleID_clean %in% (meta_both %>% filter(Continent == ct) %>% pull(SampleID_clean))) %>% 
    column_to_rownames(var = "SampleID_clean") %>% as.matrix()
  set.seed(2025)
  res[[paste0("Both ",ct)]] <- run_sparcc(mt_wk = mt_wkk, R = 1000)
}

save(res, file = paste0(temp_out,"network_20250320.Rdata"))
```

### plot
```{r}
load(file = paste0(temp_out,"network_20250320.Rdata"))
k_color = c("Fungi"="#fc8d62", "Bacteria"="#8da0cb", "Archaea"="#66c2a5")

for (nn in names(res)){
  cor_cutoff = 0.3
  p_cutoff = 0.05
  cor_sparcc = res[[nn]]$cor
  pval_sparcc = res[[nn]]$pval
  cor_sparcc[ pval_sparcc>=p_cutoff ] <- 0
  cor_tidy <- cor_gather(cor_sparcc) 
  cor_tidy <- cor_tidy %>% 
    rename(from = var1, to = var2) %>% 
    filter(abs(cor) > cor_cutoff) 
  if(nrow(cor_tidy) == 0){next}
  gdat <- tidy_graph_edge_node(cor_tidy = cor_tidy)
  gg <- tbl_graph(directed = F, nodes = gdat$node, edges = gdat$edge) 
  
  #for (ll in c('kk', 'dh', 'drl','fr','gem','circle', 'graphopt', 'grid','lgl','mds','randomly')){
  for (ll in c('fr','circle')){
    set.seed(2025)
    g <- ggraph(gg, layout = ll) +
      geom_edge_link(aes(colour = col)) +
      geom_node_point(aes(color = col_k), size = 3) +
      geom_node_text(aes(label = label3), repel = TRUE) + 
      scale_color_manual(values = k_color, name = "") +
      scale_edge_color_manual(values = c("pos"="green", "neg"="red"), name = "") + 
      labs(title = nn) + 
      theme_graph(foreground = 'steelblue', fg_text_colour = 'white', base_family = 'Helvetica')
    
    ggsave(g, file = paste0(temp_out, "sparcc_",nn,"_",ll,"_pval0.05_R6.pdf"), 
           device = pdf)
  }
}

```


## new same samples 16s and its
### use sparcc ab 0.01, pre 0.1, cor 0.3,
```{r}
abundance_cutoff <- 0.01
prevalence_cutoff <- 0.1

ft_its_wk_g_filter <- filter_ra_pre(ft_mt = ft_its_wk_g, ra_mt = ra_its_wk_g, abundance_cutoff = abundance_cutoff, prevalence_cutoff = prevalence_cutoff)
ft_its_wk_g_filter <- ft_its_wk_g_filter  %>% as.data.frame() %>% 
  rownames_to_column(var = "SampleID_clean") %>% arrange(SampleID_clean) %>% filter(SampleID_clean != "BR34.1")

ft_16s_wk_g_filter <- filter_ra_pre(ft_mt = ft_16s_wk_g, ra_mt = ra_16s_wk_g, abundance_cutoff = abundance_cutoff, prevalence_cutoff = prevalence_cutoff)
ft_16s_wk_g_filter <- ft_16s_wk_g_filter %>% as.data.frame() %>% 
  rownames_to_column(var = "SampleID_clean") %>% arrange(SampleID_clean) %>% filter(SampleID_clean != "BR34.1")

ft_16s_wk_g_filter <- ft_16s_wk_g_filter %>% filter(SampleID_clean %in% ft_its_wk_g_filter$SampleID_clean)

res <- list()

mt_wk <- ft_16s_wk_g_filter %>% column_to_rownames(var = "SampleID_clean") %>% as.matrix()
set.seed(2025)
res[["16S All"]] <- run_sparcc(mt_wk = mt_wk, R = 1000)
for(ct in c("AFRICA","AMERICA","EUROPE")){
  mt_wkk <- mt_wk %>% as.data.frame() %>% rownames_to_column(var = "SampleID_clean") %>% 
    filter(SampleID_clean %in% (meta_global %>% filter(Continent == ct) %>% pull(SampleID_clean))) %>% 
    column_to_rownames(var = "SampleID_clean") %>% as.matrix()
  set.seed(2025)
  res[[paste0("16S ",ct)]] <- run_sparcc(mt_wk = mt_wkk, R = 1000)
}  

aa <- ft_16s_wk_g_filter %>% filter(SampleID_clean %in% ft_its_wk_g_filter$SampleID_clean) %>% 
  column_to_rownames(var = "SampleID_clean")
aa <- aa[,colSums(aa)>0] %>% 
  rownames_to_column(var = "SampleID_clean")
mt_wk <- ft_its_wk_g_filter %>% 
  left_join(aa, by = "SampleID_clean")  %>% 
  column_to_rownames(var = "SampleID_clean") %>% 
  as.matrix()
set.seed(2025)
res[["Both All"]] <- run_sparcc(mt_wk = mt_wk, R = 1000)

for(ct in c("AFRICA", "AMERICA")){
  mt_wkk <- mt_wk %>% as.data.frame() %>% rownames_to_column(var = "SampleID_clean") %>% 
    filter(SampleID_clean %in% (meta_both %>% filter(Continent == ct) %>% pull(SampleID_clean))) %>% 
    column_to_rownames(var = "SampleID_clean") %>% as.matrix()
  set.seed(2025)
  res[[paste0("Both ",ct)]] <- run_sparcc(mt_wk = mt_wkk, R = 1000)
}

save(res, file = paste0(temp_out,"network_20250320_same_size.Rdata"))
```

### plot
```{r}
load(file = paste0(temp_out,"network_20250320_same_size.Rdata"))
k_color = c("Fungi"="#fc8d62", "Bacteria"="#8da0cb", "Archaea"="#66c2a5")

for (nn in names(res)){
  cor_cutoff = 0.3
  p_cutoff = 0.05
  cor_sparcc = res[[nn]]$cor
  pval_sparcc = res[[nn]]$pval
  cor_sparcc[ pval_sparcc>=p_cutoff ] <- 0
  cor_tidy <- cor_gather(cor_sparcc) 
  cor_tidy <- cor_tidy %>% 
    rename(from = var1, to = var2) %>% 
    filter(abs(cor) > cor_cutoff) 
  if(nrow(cor_tidy) == 0){next}
  gdat <- tidy_graph_edge_node(cor_tidy = cor_tidy)
  gg <- tbl_graph(directed = F, nodes = gdat$node, edges = gdat$edge) 
  

  for (ll in c('fr','circle')){
    set.seed(2025)
    g <- ggraph(gg, layout = ll) +
      geom_edge_link(aes(colour = col)) +
      geom_node_point(aes(color = col_k), size = 3) +
      geom_node_text(aes(label = label3), repel = TRUE) + 
      scale_color_manual(values = k_color, name = "") +
      scale_edge_color_manual(values = c("pos"="green", "neg"="red"), name = "") + 
      labs(title = nn) + 
      theme_graph(foreground = 'steelblue', fg_text_colour = 'white', base_family = 'Helvetica')
    
    ggsave(g, file = paste0(temp_out, "sparcc_new_same_size_",nn,"_",ll,"_pval0.05_R6.pdf"), 
           device = pdf)
  }
}

```
