==> code/09_Spain_Covid.Rmd <==
---
title: "Alpha diversity"
author: "Haipeng Sun"
date: "8/1/2022"
output: html_document
---

## library
```{r setup, message=FALSE, warning=FALSE}
library(phyloseq)
library(purrr)
library(furrr)
library(picante)
library(vegan)
library(MASS)
library(ggpubr)
library(effectsize)
library(ggtext)
library(glue)
library(usedist)
library(ape)
library(rlist)
library(ggrepel)
library(tidyverse)
temp_out <-  "~/Library/CloudStorage/Dropbox/rutgers/20220606_RUN41_milk/Data/output/COVID/"
covid_color <- c("N" = "#91CF60" , "Y" = "#FC8D59")
```

## alpha
### load data
```{r}
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/phylo_meta.Rdata")
load(file = "~/Library/CloudStorage/Dropbox/Rutgers/20220606_RUN41_milk/Data/alpha.Rdata")

dat_alpha <- left_join(alpha_boot, meta_all, by = "SampleID")
```

### alpha by covid
```{r}
dat <- dat_alpha %>% filter(Project == "Covid Spain") %>% 
  mutate(Covid = ifelse(is.na(Covid_state_mom), "N","Y")) %>% 
 dplyr::select(SampleID, Covid,Faith_PD:Pielou_evenness) %>% 
  pivot_longer(Faith_PD:Pielou_evenness, names_to = "Metrics", values_to = "value")
  

g <- ggplot(dat, aes(Covid, value)) + 
  geom_boxplot(aes(color = Covid), outlier.alpha = 0) + 
  geom_point(aes(color = Covid)) + 
  stat_compare_means() + 
  scale_x_discrete(labels = c("N"="Covid Negative", "Y"="Covid Positive")) + 
  scale_color_manual(values = covid_color) + 
  labs(x="", y="") + 
  facet_wrap(~Metrics, nrow = 2, scales = "free_y") +
  theme_bw() + theme(#aspect.ratio = 0.8,
                     legend.position="none",
                     panel.background = element_rect(fill = NA),
                     strip.text = element_text(size=12,color="black",face='bold'),
                     axis.title = element_text(size=12,color="black",face='bold'),
                     axis.text = element_text(size=12,color="black",face='bold'),
                     axis.text.x = element_text(size=12,color="black",face='bold',angle=0),
                     axis.text.y = element_text(size=12,color="black",face='bold'),
                     legend.text = element_text(size=12,color="black",face='bold'),
                     legend.title = element_text(size=12,color="black",face='bold'),
                     title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out, "alpha_by_covid.pdf"), device = cairo_pdf, 
         plot = g, width = 9, height = 7, units = "in", dpi = 300) 


```

## beta 
### load data
```{r}
load(file = "~/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_uu.RData")
load(file = "~/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_wu.RData")
load(file = "~/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_bray.RData")
load(file = "~/Dropbox/Rutgers/20220606_RUN41_milk/Data/beta_DM_jaccard.RData")

calcOmega2_dm <- function(x) {
    # Computes the effect size omega^2 parital using the output from adonis test
    # Args:
    #   x: adonis output
    # Returns:
    #   A dataframe with calculated effect size omega^2 partial added to original dataframe
    require(dplyr)
    N_t = x["Total", ]$Df + 1
    MSe = x["Residual", ]$SumOfSqs/x["Residual", ]$Df
    out = x %>% as.data.frame %>% rownames_to_column(var = "Variable") %>% mutate(Omega2_partial = ifelse(is.na(`F`), NA, (SumOfSqs - Df * MSe)/(SumOfSqs + (N_t - Df) * MSe)))
    return(out)
}

```

### PCoa
```{r}
meta <- phylo_decontam_clean_rf@sam_data@.Data %>% as.data.frame()
names(meta) <- phylo_decontam_clean_rf@sam_data@names
meta$SampleID <- phylo_decontam_clean_rf@sam_data@row.names

meta <- meta %>% filter(Project == "Covid Spain") %>% 
  mutate(Covid = ifelse(is.na(Covid_state_mom), "N","Y"))

for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta = meta %>% filter(SampleID %in% labels(wdist))
  wdist <- dist_subset(wdist, wmeta$SampleID) 
  #adonis
  per <- adonis2(wdist ~ Covid, 
                 data = wmeta, by = "margin")
  res = calcOmega2_dm(per)
  write.csv(res, file = paste0(temp_out,"adonis_test_", dd, "_Covid.csv"), quote = F, row.names = F)
  #PCoA
  wrda <- dbrda(wdist ~ 1, sqrt.dist = TRUE)
  wrda_eig <- eigenvals(wrda)
  varPC1 <- round(wrda_eig["MDS1"]/sum(wrda_eig)*100,2)
  varPC2 <- round(wrda_eig["MDS2"]/sum(wrda_eig)*100,2)
  x_lab <- paste0("PC1 (", varPC1, "%)")
  y_lab <- paste0("PC2 (", varPC2, "%)")
  wdat_pcoa <- merge(scores(wrda, display = "sites", choices = c(1,2)),
                    wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = MDS1, PC2 = MDS2)
  ## Covid
  wrda_fit <- envfit(wrda ~ Covid, data = wmeta)
  wdat_center <- wrda_fit$factors$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "Covid") %>%
    transmute(Covid = str_remove(Covid,"Covid"),PC1 = MDS1, PC2 = MDS2) 
  wdat <- wdat_pcoa %>% left_join(wdat_center %>% transmute(PC1.c = PC1, PC2.c = PC2, Covid), by = "Covid")

  g <- ggplot(wdat, aes(x = PC1, y = PC2, color = Covid)) + 
      geom_point(size = 2) + 
      scale_color_manual(values = covid_color, 
                         labels = c("N"="Covid Negative", "Y"="Covid Positive"),
                         name = "") + 
      labs(x = x_lab, y = y_lab, title = "") +
      theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))    
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_covid.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)
  
  g <- ggplot(wdat_center, aes(x = PC1, y = PC2, color = Covid)) + 
    geom_point(data = wdat, aes(x = PC1, y = PC2, color = Covid), size = 1, alpha = 0.8) + 
    geom_segment(data = wdat, aes(x = PC1, y = PC2, xend = PC1.c, yend = PC2.c), size = 0.2, alpha = 0.8) + 
    geom_point(size = 4,shape = 21, color = "black", aes(fill = Covid)) + 
    geom_text_repel(aes(label = Covid), color = "black") + 
    scale_color_manual(values = covid_color, aesthetics = c("color","fill")) + 
    labs(x = x_lab, y = y_lab, title = "") +
    theme_bw() + theme(#aspect.ratio = 1,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))    
  ggsave(filename = paste0(temp_out,"PCoA_",dd,"_Covid_center.pdf"), device = cairo_pdf, 
           plot = g, width = 7, height = 5, units = "in", dpi = 300)  
}
```


## taxa ancom
### data
```{r}
load(file = "~/Dropbox/Rutgers/20220606_RUN41_milk/Data/phylo_meta.Rdata")
meta <- meta_all %>% filter(Project == "Covid Spain") %>% 
  mutate(Covid = ifelse(is.na(Covid_state_mom), "N","Y"))

phylo_covid <- prune_samples(meta$SampleID, phylo_decontam_clean_rf)
new_sam_data <- sample_data(meta %>% column_to_rownames(var = "SampleID"))
sample_data(phylo_covid) <- new_sam_data

phylo_s <- tax_glom(phylo_covid, taxrank = "Species", NArm = T)

```


### ancom
```{r}
library(ANCOMBC)
library(mia)



taxa_tmp <- tax_table(phylo_s@tax_table) %>% as.matrix() %>% as.data.frame() %>% 
  rownames_to_column(var = "id") %>% 
  mutate(Species = str_c(Species,";",id)) %>% 
  column_to_rownames(var = "id") %>% as.matrix()
taxa_tmp <- tax_table(taxa_tmp)

tax_table(phylo_s) <- taxa_tmp
out <- ancombc2(data = phylo_s, fix_formula = "Covid", tax_level = "Species")
  
taxa_tmp <- tax_table(phylo_covid@tax_table) %>% as.matrix() %>% as.data.frame() %>% 
  rownames_to_column(var = "id") %>% 
  mutate(Species = id) %>% 
  column_to_rownames(var = "id") %>% as.matrix()
taxa_tmp <- tax_table(taxa_tmp)
tax_table(phylo_covid) <- taxa_tmp

out_asv <- ancombc2(data = phylo_covid, fix_formula = "Covid", tax_level = "Species")  
  


```
### maaslin2
```{r}
library(Maaslin2)
do_maaslin2 <- function(meta, sample_id, phylo_rf, fixed, random, ref, output, 
                       min_prevalence = 0.1, min_abundance = 0.0, max_significance = 0.1,
                       normalization = "TSS", transform = "LOG",analysis_method = "LM",
                       plot_heatmap	= FALSE, plot_scatter = FALSE, cores = 1){
  ft <- otu_table(phylo_rf@otu_table) %>% as.matrix() %>% as.data.frame() %>% rownames_to_column(var = "id")
  wft <- ft %>% dplyr::select(id,  meta %>% pull(sample_id)) %>% column_to_rownames(var = "id") %>% t()
  wmeta <- meta %>% column_to_rownames(var = sample_id)
  taxrank <- "ASV"
  fit <- Maaslin2(
    input_data = wft, 
    input_metadata = wmeta, 
    normalization = normalization,
    transform = transform,
    analysis_method = analysis_method,
    output = paste0(output,"_",taxrank),
    fixed_effects = fixed,
    reference = ref,
    random_effects = random,
    min_prevalence = min_prevalence,
    min_abundance = min_abundance,
    max_significance = max_significance,
    plot_heatmap = plot_heatmap,
    plot_scatter = plot_scatter,
    cores = cores
  )
  for (taxrank in c("Species","Genus","Family","Order","Class","Phylum")){
  phylo_glom <- tax_glom(phylo_rf, taxrank = taxrank, NArm = F)
  ft_glom <- otu_table(phylo_glom@otu_table) %>% as.matrix() %>% as.data.frame() %>% rownames_to_column(var = "id")
  wft <- ft_glom %>% dplyr::select(id, meta %>% pull(sample_id)) %>% column_to_rownames(var = "id") %>% t()
  fit <- Maaslin2(
      input_data = wft, 
      input_metadata = wmeta, 
      normalization = normalization,
      transform = transform,
      analysis_method = analysis_method,
      output = paste0(output,"_",taxrank),
      fixed_effects = fixed,
      reference = ref,
      random_effects = random,
      min_prevalence = min_prevalence,
      min_abundance = min_abundance,
      max_significance = max_significance,
      plot_heatmap = plot_heatmap,
      plot_scatter = plot_scatter,
      cores = cores
      )
  }
}
summarise_maaslin2 <- function(path, group, taxa){
  res <- matrix(data = NA, nrow = 1, ncol = 20)
  res <- as.data.frame(res)
  names(res) <-  c("feature","metadata","value","coef","stderr","N","N.not.0","pval","qval","Group",
                  "id","Kingdom","Phylum","Class","Order","Family","Genus","Species","ASV", "level")
  for (taxrank in c("ASV", "Species", "Genus", "Family", "Order", "Class", "Phylum")){
    a <- read.table(file = paste0(path ,"_", taxrank, "/significant_results.tsv"),
                  header = T, sep = "\t", stringsAsFactors = F)
    if (nrow(a) == 0) {next}
    a <- a %>% mutate(Group = group, id = str_remove(feature,"^X"),feature = str_remove(feature,"^X")) %>% left_join(taxa, by = "id") %>% mutate(level = taxrank)
    res <- rbind(res,a)
  }
  res <- res %>% filter(!is.na(feature))
  res
}
output <- paste0(temp_out, "maaslin2/")
fit <- do_maaslin2(meta = meta, sample_id = "SampleID", phylo_rf = phylo_covid, 
                  fixed = c("Covid"), 
                  random = NULL, 
                  ref = c("N"), 
                  output = paste0(output,"All"),
                  cores = 4
                  )
res <- summarise_maaslin2(path = paste0(output,"All"), group = "all", taxa %>% dplyr::select(-taxa_long))


```

### lesfe
```{r}
ft <- otu_table(phylo_s@otu_table) %>% as.data.frame()
ft <- ft[rowSums(ft)>0,]
ft <- ft %>% rownames_to_column(var = "id")

taxa_tmp <- tax_table(phylo_covid@tax_table) %>% as.matrix() %>% as.data.frame() %>% 
  rownames_to_column(var = "id") %>%  replace(is.na(.), "unknown") %>% 
  mutate(Kingdom = str_remove(Kingdom, "^d__")) %>% 
  mutate(taxa_long =str_c(Kingdom, Phylum, Class, Order, Family, Genus, Species, sep = "|"))

ft <-  ft %>% left_join(taxa_tmp %>% dplyr::select(id, taxa_long), by = "id") %>% 
  mutate(id = taxa_long) %>% select(-taxa_long) 

a <- meta %>% dplyr::select(SampleID, Covid) %>% t()

write.table(a, file = paste0(temp_out,"lefse_meta.txt"),
            append = F, sep = "\t", quote = F, row.names = F )

write.table(ft, file = paste0(temp_out,"lefse_otu.txt"),
            append = F, sep = "\t", quote = F, row.names = F )

```
